Split the trail data array into mountain-bike-trails.data.ts
mountain-bike-trails.ts was 3,261 lines — ~3,000 of them a single array
literal that buried the interfaces and region logic. Move the array (and
the trailColor helper that only the array uses) into a new
mountain-bike-trails.data.ts; the original keeps the MountainBikeTrail /
ElevationProfile interfaces, layer-id constants, REGION_MAP, and regionFor,
and re-exports the array. The data file's only import from the logic file
is the MountainBikeTrail type, which is erased at compile time — no runtime
circular dependency. Consumers importing from @/data/geo_data are
unaffected.

Update the trail scripts to match:
- Both point at mountain-bike-trails.data.ts (add_trail_bounds.py was also
  pointing at the wrong file, geo_data.ts, which never held the array).
- add_trail_elevation.py reads the Mapbox token from NEXT_PUBLIC_MAPBOX_TOKEN
  (env or .env.local) — Phase 1 moved the token out of map.config.ts, which
  had broken the script's regex-based extraction.

docs/DATA.md updated to point curators at the data file.
