+++ scripts/add_trail_elevation.py
@@ -6,7 +6,7 @@ from Terrain-RGB tiles. No browser extraction needed.
 
 Output:
   - public/data/elevation/{slug}.json  — per-trail elevation profiles
-  - src/data/geo_data.ts               — summary stats (gain, loss, min, max)
+  - src/data/mountain-bike-trails.data.ts  — summary stats (gain, loss, min, max)
 
 Usage: python scripts/add_trail_elevation.py
 Requires: pip install -r scripts/requirements.txt
@@ -67,7 +67,7 @@ SAMPLE_STEP_FT = 25
 METERS_TO_FEET = 3.28084
 EARTH_RADIUS_FT = 20902231.0
 OUTPUT_DIR = 'public/data/elevation'
-GEO_DATA_PATH = 'src/data/mountain-bike-trails.ts'
+GEO_DATA_PATH = 'src/data/mountain-bike-trails.data.ts'
 TILE_CACHE_DIR = 'scripts/.tile_cache'
 
 # --- Caches ---
@@ -539,7 +539,7 @@ def extract_all_trails(zoom, bbox):
 
 
 def get_known_trails():
-    """Extract trail names and defaultBounds from geo_data.ts."""
+    """Extract trail names and defaultBounds from mountain-bike-trails.data.ts."""
     with open(GEO_DATA_PATH) as f:
         content = f.read()
 
@@ -624,11 +624,11 @@ def slugify(name):
 
 
 # ============================================================
-# geo_data.ts Updater
+# mountain-bike-trails.data.ts Updater
 # ============================================================
 
 def update_geo_data(trail_data):
-    """Update geo_data.ts with distance and elevation summary stats.
+    """Update mountain-bike-trails.data.ts with distance and elevation summary stats.
 
     Uses a line-by-line approach to avoid regex crossing trail boundaries.
     """
@@ -777,10 +777,10 @@ def main():
     print(f"\nDone! {len(trail_results)} profiles in {time.time() - start_time:.1f}s")
     print(f"Terrain tiles: {len(terrain_cache)}, MVT tiles: {len(mvt_cache)}")
 
-    # Step 3: Update mountain-bike-trails.ts
-    print("\nUpdating mountain-bike-trails.ts...")
+    # Step 3: Update mountain-bike-trails.data.ts
+    print("\nUpdating mountain-bike-trails.data.ts...")
     updated = update_geo_data(trail_results)
-    print(f"Updated {updated} trails in mountain-bike-trails.ts")
+    print(f"Updated {updated} trails in mountain-bike-trails.data.ts")
 
 
 if __name__ == '__main__':
