+++ src/data/mountain-bike-trails.data.ts
@@ -1,0 +1,3191 @@
+import { faRoute, faMountain } from '@fortawesome/free-solid-svg-icons';
+import type { MountainBikeTrail } from './mountain-bike-trails';
+
+// Trail colors by difficulty rating
+const TRAIL_COLOR_EASY = '#16A34A';
+const TRAIL_COLOR_INTERMEDIATE = '#2563EB';
+const TRAIL_COLOR_ADVANCED = '#374151';
+const TRAIL_COLOR_EXPERT = '#000000';
+const TRAIL_COLOR_UNRATED = '#6B7280';
+const TRAIL_COLOR_GREENWAY = '#059669';
+
+function trailColor(rating: string, isGreenway: boolean): string {
+  if (isGreenway) return TRAIL_COLOR_GREENWAY;
+  if (rating === 'easy') return TRAIL_COLOR_EASY;
+  if (rating === 'intermediate') return TRAIL_COLOR_INTERMEDIATE;
+  if (rating === 'advanced') return TRAIL_COLOR_ADVANCED;
+  if (rating === 'expert') return TRAIL_COLOR_EXPERT;
+  return TRAIL_COLOR_UNRATED;
+}
+
+export const mountainBikeTrails: MountainBikeTrail[] = [
+  // Bauxite Ridge/White Oak Mountain
+  {
+    trailName: 'Big Forest',
+    displayName: 'Big Forest',
+    recArea: 'White Oak Mountain',
+    rating: '',
+    color: trailColor('', false),
+    distance: 0.7,
+    elevationGain: 224,
+    elevationLoss: 243,
+    elevationMin: 973,
+    elevationMax: 1075,
+    defaultBounds: [-85.041046, 35.038851, -85.037956, 35.045457],
+    icon: faMountain,
+  },
+  {
+    trailName: 'Crashing Creek',
+    displayName: 'Crashing Creek',
+    recArea: 'White Oak Mountain',
+    rating: '',
+    color: trailColor('', false),
+    distance: 0.4,
+    elevationGain: 8,
+    elevationLoss: 49,
+    elevationMin: 820,
+    elevationMax: 868,
+    defaultBounds: [-85.0383, 35.020578, -85.031433, 35.021843],
+    icon: faMountain,
+  },
+  {
+    trailName: 'Forest Lane',
+    displayName: 'Forest Lane',
+    recArea: 'White Oak Mountain',
+    rating: 'easy',
+    color: trailColor('easy', false),
+    distance: 0.8,
+    elevationGain: 286,
+    elevationLoss: 78,
+    elevationMin: 875,
+    elevationMax: 1083,
+    defaultBounds: [-85.043278, 35.042365, -85.039673, 35.045457],
+    icon: faMountain,
+  },
+  {
+    trailName: 'Murky Marsh',
+    displayName: 'Murky Marsh',
+    recArea: 'White Oak Mountain',
+    rating: 'intermediate',
+    color: trailColor('intermediate', false),
+    distance: 4.3,
+    elevationGain: 650,
+    elevationLoss: 788,
+    elevationMin: 840,
+    elevationMax: 1025,
+    defaultBounds: [-85.040703, 35.018329, -85.035038, 35.034916],
+    icon: faMountain,
+  },
+  {
+    trailName: 'Oak Hole Lane',
+    displayName: 'Oak Hole Lane',
+    recArea: 'White Oak Mountain',
+    rating: 'easy',
+    color: trailColor('easy', false),
+    distance: 1.0,
+    elevationGain: 219,
+    elevationLoss: 219,
+    elevationMin: 1012,
+    elevationMax: 1089,
+    defaultBounds: [-85.041561, 35.039414, -85.0383, 35.045035],
+    icon: faMountain,
+  },
+  {
+    trailName: "Pogo's Place",
+    displayName: "Pogo's Place",
+    recArea: 'White Oak Mountain',
+    rating: 'intermediate',
+    color: trailColor('intermediate', false),
+    distance: 1.2,
+    elevationGain: 246,
+    elevationLoss: 237,
+    elevationMin: 999,
+    elevationMax: 1064,
+    defaultBounds: [-85.041389, 35.02339, -85.0383, 35.029153],
+    icon: faMountain,
+  },
+  {
+    trailName: 'Rolling Ridge',
+    displayName: 'Rolling Ridge',
+    recArea: 'White Oak Mountain',
+    rating: '',
+    color: trailColor('', false),
+    distance: 1.5,
+    elevationGain: 305,
+    elevationLoss: 419,
+    elevationMin: 915,
+    elevationMax: 1054,
+    defaultBounds: [-85.042419, 35.029294, -85.036755, 35.039414],
+    icon: faMountain,
+  },
+  // Booker T Washington SP
+  {
+    trailName: 'Inner Loop 1',
+    displayName: 'Inner Loop 1',
+    recArea: 'Booker T Washington State Park',
+    rating: '',
+    color: trailColor('', false),
+    distance: 1.0,
+    elevationGain: 164,
+    elevationLoss: 180,
+    elevationMin: 702,
+    elevationMax: 796,
+    defaultBounds: [-85.177345, 35.114292, -85.172882, 35.118926],
+    icon: faMountain,
+  },
+  {
+    trailName: 'Inner Loop 2',
+    displayName: 'Inner Loop 2',
+    recArea: 'Booker T Washington State Park',
+    rating: '',
+    color: trailColor('', false),
+    distance: 2.0,
+    elevationGain: 279,
+    elevationLoss: 286,
+    elevationMin: 693,
+    elevationMax: 776,
+    defaultBounds: [-85.175285, 35.116819, -85.169277, 35.12398],
+    icon: faMountain,
+  },
+  {
+    trailName: 'Outer Bike Trail',
+    displayName: 'Outer Bike Trail',
+    recArea: 'Booker T Washington State Park',
+    rating: '',
+    color: trailColor('', false),
+    distance: 4.6,
+    elevationGain: 929,
+    elevationLoss: 921,
+    elevationMin: 683,
+    elevationMax: 812,
+    defaultBounds: [-85.178547, 35.112186, -85.167046, 35.124121],
+    icon: faMountain,
+  },
+  // Brainard Levee Trail
+  {
+    trailName: 'Brainard Levee Trail',
+    displayName: 'Brainard Levee Trail',
+    recArea: 'Chattanooga Greenways',
+    rating: '',
+    color: trailColor('', true),
+    distance: 0.8,
+    elevationGain: 157,
+    elevationLoss: 238,
+    elevationMin: 638,
+    elevationMax: 745,
+    defaultBounds: [-85.215626, 35.059369, -85.211334, 35.0678],
+    icon: faRoute,
+  },
+  // Camp Jordan
+  {
+    trailName: 'Camp Jordan Loop',
+    displayName: 'Camp Jordan Loop',
+    recArea: 'Chattanooga Greenways',
+    rating: '',
+    color: trailColor('', true),
+    distance: 6.2,
+    elevationGain: 209,
+    elevationLoss: 193,
+    elevationMin: 649,
+    elevationMax: 690,
+    defaultBounds: [-85.218716, 34.997519, -85.19125, 35.044333],
+    icon: faRoute,
+  },
+  // Chattahoochee National Forest
+  {
+    trailName: 'Hudon Loop',
+    displayName: 'Hudon Loop',
+    recArea: 'Houston Valley',
+    rating: '',
+    color: trailColor('', true),
+    distance: 0.3,
+    elevationGain: 34,
+    elevationLoss: 34,
+    elevationMin: 940,
+    elevationMax: 972,
+    defaultBounds: [-85.121212, 34.787444, -85.120354, 34.788995],
+    icon: faRoute,
+  },
+  {
+    trailName: 'Hudon Spur',
+    displayName: 'Hudon Spur',
+    recArea: 'Houston Valley',
+    rating: '',
+    color: trailColor('', true),
+    distance: 0.6,
+    elevationGain: 52,
+    elevationLoss: 54,
+    elevationMin: 926,
+    elevationMax: 974,
+    defaultBounds: [-85.129452, 34.787444, -85.120354, 34.788431],
+    icon: faRoute,
+  },
+  // Chattanooga Greenways
+  {
+    trailName: 'Moccassin Bend Greenway',
+    displayName: 'Moccassin Bend Greenway',
+    recArea: 'Chattanooga Greenways',
+    rating: '',
+    color: trailColor('', true),
+    distance: 1.1,
+    elevationGain: 87,
+    elevationLoss: 83,
+    elevationMin: 634,
+    elevationMax: 676,
+    defaultBounds: [-85.313473, 35.056418, -85.306606, 35.062741],
+    icon: faRoute,
+  },
+  // Chickamauga Battlefield National Military Park
+  {
+    trailName: 'BLUE Trail',
+    displayName: 'BLUE Trail',
+    recArea: 'Chickamauga Battlefield National Military Park',
+    rating: '',
+    color: trailColor('', false),
+    distance: 8.6,
+    elevationGain: 491,
+    elevationLoss: 480,
+    elevationMin: 688,
+    elevationMax: 795,
+    defaultBounds: [-85.273819, 34.895506, -85.227814, 34.941955],
+    icon: faMountain,
+  },
+  {
+    trailName: 'CLOSED Road',
+    displayName: 'CLOSED Road',
+    recArea: 'Chickamauga Battlefield National Military Park',
+    rating: '',
+    color: trailColor('', false),
+    distance: 0.2,
+    elevationGain: 46,
+    elevationLoss: 9,
+    elevationMin: 797,
+    elevationMax: 842,
+    defaultBounds: [-85.271587, 34.92929, -85.268154, 34.930275],
+    icon: faMountain,
+  },
+  {
+    trailName: 'RED Trail',
+    displayName: 'RED Trail',
+    recArea: 'Chickamauga Battlefield National Military Park',
+    rating: '',
+    color: trailColor('', false),
+    distance: 9.3,
+    elevationGain: 698,
+    elevationLoss: 704,
+    elevationMin: 706,
+    elevationMax: 916,
+    defaultBounds: [-85.278282, 34.90888, -85.234509, 34.939141],
+    icon: faMountain,
+  },
+  {
+    trailName: 'WHITE Trail',
+    displayName: 'WHITE Trail',
+    recArea: 'Chickamauga Battlefield National Military Park',
+    rating: '',
+    color: trailColor('', false),
+    distance: 2.2,
+    elevationGain: 32,
+    elevationLoss: 81,
+    elevationMin: 740,
+    elevationMax: 790,
+    defaultBounds: [-85.259914, 34.909443, -85.25425, 34.932949],
+    icon: faMountain,
+  },
+  // Cleveland/Bradley County Greenway
+  {
+    trailName: 'Mouse Creek Greenway Phase 1',
+    displayName: 'Mouse Creek Greenway Phase 1',
+    recArea: 'Cleveland',
+    rating: '',
+    color: trailColor('', true),
+    distance: 0.6,
+    elevationGain: 15,
+    elevationLoss: 18,
+    elevationMin: 796,
+    elevationMax: 804,
+    defaultBounds: [-84.876938, 35.175211, -84.87299, 35.182087],
+    icon: faRoute,
+  },
+  {
+    trailName: 'Mouse Creek Greenway Phase 2',
+    displayName: 'Mouse Creek Greenway Phase 2',
+    recArea: 'Cleveland',
+    rating: '',
+    color: trailColor('', true),
+    distance: 0.8,
+    elevationGain: 23,
+    elevationLoss: 29,
+    elevationMin: 790,
+    elevationMax: 799,
+    defaultBounds: [-84.87299, 35.182087, -84.865093, 35.191767],
+    icon: faRoute,
+  },
+  {
+    trailName: 'Mouse Creek Greenway Phase 3',
+    displayName: 'Mouse Creek Greenway Phase 3',
+    recArea: 'Cleveland',
+    rating: '',
+    color: trailColor('', true),
+    distance: 1.1,
+    elevationGain: 101,
+    elevationLoss: 66,
+    elevationMin: 803,
+    elevationMax: 873,
+    defaultBounds: [-84.879856, 35.166511, -84.876938, 35.175211],
+    icon: faRoute,
+  },
+  {
+    trailName: 'Mouse Creek Greenway Phase 4',
+    displayName: 'Mouse Creek Greenway Phase 4',
+    recArea: 'Cleveland',
+    rating: '',
+    color: trailColor('', true),
+    distance: 1.2,
+    elevationGain: 58,
+    elevationLoss: 46,
+    elevationMin: 757,
+    elevationMax: 776,
+    defaultBounds: [-84.85548, 35.200885, -84.852047, 35.211825],
+    icon: faRoute,
+  },
+  {
+    trailName: 'Mouse Creek Greenway Phase 5',
+    displayName: 'Mouse Creek Greenway Phase 5',
+    recArea: 'Cleveland',
+    rating: '',
+    color: trailColor('', true),
+    distance: 1.7,
+    elevationGain: 84,
+    elevationLoss: 138,
+    elevationMin: 762,
+    elevationMax: 824,
+    defaultBounds: [-84.865093, 35.191767, -84.852562, 35.201727],
+    icon: faRoute,
+  },
+  {
+    trailName: 'Mouse Creek Greenway Phase 6',
+    displayName: 'Mouse Creek Greenway Phase 6',
+    recArea: 'Cleveland',
+    rating: '',
+    color: trailColor('', true),
+    distance: 0.6,
+    elevationGain: 34,
+    elevationLoss: 25,
+    elevationMin: 811,
+    elevationMax: 841,
+    defaultBounds: [-84.882603, 35.162863, -84.879856, 35.168195],
+    icon: faRoute,
+  },
+  // Cloudland Canyon SP
+  {
+    trailName: 'Barkeater Spur',
+    displayName: 'Barkeater Spur',
+    recArea: 'Cloudland Canyon State Park',
+    rating: 'intermediate',
+    color: trailColor('intermediate', false),
+    distance: 0.1,
+    elevationGain: 0,
+    elevationLoss: 67,
+    elevationMin: 1808,
+    elevationMax: 1875,
+    defaultBounds: [-85.44445, 34.8317, -85.442905, 34.832968],
+    icon: faMountain,
+  },
+  {
+    trailName: 'Cloudland Canyon Connector',
+    displayName: 'Cloudland Canyon Connector',
+    recArea: 'Cloudland Canyon State Park',
+    rating: 'easy',
+    color: trailColor('easy', false),
+    distance: 9.6,
+    elevationGain: 1771,
+    elevationLoss: 1737,
+    elevationMin: 1621,
+    elevationMax: 2048,
+    defaultBounds: [-85.480156, 34.817045, -85.413208, 34.862553],
+    icon: faMountain,
+  },
+  {
+    trailName: 'Pathkiller',
+    displayName: 'Pathkiller',
+    recArea: 'Cloudland Canyon State Park',
+    rating: '',
+    color: trailColor('', false),
+    distance: 1.0,
+    elevationGain: 82,
+    elevationLoss: 85,
+    elevationMin: 1868,
+    elevationMax: 1898,
+    defaultBounds: [-85.482731, 34.818172, -85.479298, 34.823245],
+    icon: faMountain,
+  },
+  // Collegedale Greenways
+  {
+    trailName: 'Collegedale Greenway',
+    displayName: 'Collegedale Greenway',
+    recArea: 'Collegedale Greenways',
+    rating: 'easy',
+    color: trailColor('easy', true),
+    distance: 2.3,
+    elevationGain: 134,
+    elevationLoss: 128,
+    elevationMin: 759,
+    elevationMax: 791,
+    defaultBounds: [-85.071259, 35.052905, -85.049458, 35.062741],
+    icon: faRoute,
+  },
+  {
+    trailName: 'Wolftever Creek Greenway Phase 3',
+    displayName: 'Wolftever Creek Greenway Phase 3',
+    recArea: 'Collegedale Greenways',
+    rating: '',
+    color: trailColor('', true),
+    distance: 0.2,
+    elevationGain: 15,
+    elevationLoss: 3,
+    elevationMin: 779,
+    elevationMax: 791,
+    defaultBounds: [-85.049629, 35.053467, -85.046539, 35.053608],
+    icon: faRoute,
+  },
+  {
+    trailName: 'Wolftever Creek Greenway Phase 4',
+    displayName: 'Wolftever Creek Greenway Phase 4',
+    recArea: 'Collegedale Greenways',
+    rating: '',
+    color: trailColor('', true),
+    distance: 1.1,
+    elevationGain: 231,
+    elevationLoss: 97,
+    elevationMin: 778,
+    elevationMax: 924,
+    defaultBounds: [-85.045338, 35.053748, -85.03933, 35.062179],
+    icon: faRoute,
+  },
+  {
+    trailName: 'Wolftever Creek Greenway Phase 5',
+    displayName: 'Wolftever Creek Greenway Phase 5',
+    recArea: 'Collegedale Greenways',
+    rating: '',
+    color: trailColor('', true),
+    distance: 0.3,
+    elevationGain: 67,
+    elevationLoss: 103,
+    elevationMin: 777,
+    elevationMax: 827,
+    defaultBounds: [-85.060787, 35.058105, -85.057697, 35.061898],
+    icon: faRoute,
+  },
+  {
+    trailName: 'Wolftever Greenway',
+    displayName: 'Wolftever Greenway',
+    recArea: 'Collegedale Greenways',
+    rating: '',
+    color: trailColor('', true),
+    distance: 0.3,
+    elevationGain: 3,
+    elevationLoss: 3,
+    elevationMin: 759,
+    elevationMax: 760,
+    defaultBounds: [-85.069542, 35.06232, -85.068684, 35.063585],
+    icon: faRoute,
+  },
+  // Durham (LLLT)
+  {
+    trailName: '4 Tons',
+    displayName: '4 Tons',
+    recArea: 'Durham',
+    rating: 'advanced',
+    color: trailColor('advanced', false),
+    distance: 1.6,
+    elevationGain: 462,
+    elevationLoss: 252,
+    elevationMin: 1555,
+    elevationMax: 1824,
+    defaultBounds: [-85.423336, 34.874243, -85.416641, 34.878327],
+    icon: faMountain,
+  },
+  {
+    trailName: '5 Points Expressway',
+    displayName: '5 Points Expressway',
+    recArea: 'Durham',
+    rating: 'easy',
+    color: trailColor('easy', false),
+    distance: 0.6,
+    elevationGain: 88,
+    elevationLoss: 127,
+    elevationMin: 1852,
+    elevationMax: 1917,
+    defaultBounds: [-85.422649, 34.861848, -85.420933, 34.867482],
+    icon: faMountain,
+  },
+  {
+    trailName: 'Chalkeye',
+    displayName: 'Chalkeye',
+    recArea: 'Durham',
+    rating: 'easy',
+    color: trailColor('easy', false),
+    distance: 1.3,
+    elevationGain: 183,
+    elevationLoss: 219,
+    elevationMin: 1747,
+    elevationMax: 1852,
+    defaultBounds: [-85.422306, 34.865792, -85.4175, 34.874102],
+    icon: faMountain,
+  },
+  {
+    trailName: 'Cut Thru',
+    displayName: 'Cut Thru',
+    recArea: 'Durham',
+    rating: 'easy',
+    color: trailColor('easy', false),
+    distance: 0.3,
+    elevationGain: 27,
+    elevationLoss: 79,
+    elevationMin: 1772,
+    elevationMax: 1828,
+    defaultBounds: [-85.422649, 34.872271, -85.418701, 34.873679],
+    icon: faMountain,
+  },
+  {
+    trailName: 'DC&C',
+    displayName: 'DC&C',
+    recArea: 'Durham',
+    rating: 'easy',
+    color: trailColor('easy', false),
+    distance: 2.9,
+    elevationGain: 503,
+    elevationLoss: 544,
+    elevationMin: 1777,
+    elevationMax: 1918,
+    defaultBounds: [-85.432091, 34.85875, -85.41853, 34.871426],
+    icon: faMountain,
+  },
+  {
+    trailName: 'Dingey',
+    displayName: 'Dingey',
+    recArea: 'Durham',
+    rating: 'easy',
+    color: trailColor('easy', false),
+    distance: 0.3,
+    elevationGain: 34,
+    elevationLoss: 21,
+    elevationMin: 1808,
+    elevationMax: 1830,
+    defaultBounds: [-85.432262, 34.866215, -85.430546, 34.868327],
+    icon: faMountain,
+  },
+  {
+    trailName: 'Dinkey',
+    displayName: 'Dinkey',
+    recArea: 'Durham',
+    rating: 'easy',
+    color: trailColor('easy', false),
+    distance: 0.6,
+    elevationGain: 64,
+    elevationLoss: 75,
+    elevationMin: 1754,
+    elevationMax: 1802,
+    defaultBounds: [-85.425396, 34.870581, -85.422478, 34.874243],
+    icon: faMountain,
+  },
+  {
+    trailName: 'East Fugett',
+    displayName: 'East Fugett',
+    recArea: 'Durham',
+    rating: 'advanced',
+    color: trailColor('advanced', false),
+    distance: 0.5,
+    elevationGain: 290,
+    elevationLoss: 49,
+    elevationMin: 1569,
+    elevationMax: 1811,
+    defaultBounds: [-85.421963, 34.872693, -85.418186, 34.875228],
+    icon: faMountain,
+  },
+  {
+    trailName: 'John Henry',
+    displayName: 'John Henry',
+    recArea: 'Durham',
+    rating: 'easy',
+    color: trailColor('easy', false),
+    distance: 0.5,
+    elevationGain: 86,
+    elevationLoss: 47,
+    elevationMin: 1803,
+    elevationMax: 1852,
+    defaultBounds: [-85.423679, 34.865792, -85.418873, 34.87044],
+    icon: faMountain,
+  },
+  {
+    trailName: 'Marion 21',
+    displayName: 'Marion 21',
+    recArea: 'Durham',
+    rating: 'intermediate',
+    color: trailColor('intermediate', false),
+    distance: 2.8,
+    elevationGain: 663,
+    elevationLoss: 603,
+    elevationMin: 1818,
+    elevationMax: 1935,
+    defaultBounds: [-85.429001, 34.866074, -85.420074, 34.869454],
+    icon: faMountain,
+  },
+  {
+    trailName: 'Plug',
+    displayName: 'Plug',
+    recArea: 'Durham',
+    rating: 'easy',
+    color: trailColor('easy', false),
+    distance: 0.5,
+    elevationGain: 134,
+    elevationLoss: 65,
+    elevationMin: 1770,
+    elevationMax: 1851,
+    defaultBounds: [-85.424023, 34.869032, -85.422306, 34.873257],
+    icon: faMountain,
+  },
+  {
+    trailName: 'Steam Shovel',
+    displayName: 'Steam Shovel',
+    recArea: 'Durham',
+    rating: 'advanced',
+    color: trailColor('advanced', false),
+    distance: 2.2,
+    elevationGain: 512,
+    elevationLoss: 556,
+    elevationMin: 1803,
+    elevationMax: 1862,
+    defaultBounds: [-85.431919, 34.863257, -85.419044, 34.870158],
+    icon: faMountain,
+  },
+  {
+    trailName: 'West Fugett',
+    displayName: 'West Fugett',
+    recArea: 'Durham',
+    rating: 'advanced',
+    color: trailColor('advanced', false),
+    distance: 2.0,
+    elevationGain: 562,
+    elevationLoss: 753,
+    elevationMin: 1539,
+    elevationMax: 1822,
+    defaultBounds: [-85.432606, 34.866356, -85.422821, 34.875933],
+    icon: faMountain,
+  },
+  {
+    trailName: 'Williams Cut',
+    displayName: 'Williams Cut',
+    recArea: 'Durham',
+    rating: 'intermediate',
+    color: trailColor('intermediate', false),
+    distance: 0.9,
+    elevationGain: 276,
+    elevationLoss: 183,
+    elevationMin: 1803,
+    elevationMax: 1913,
+    defaultBounds: [-85.432091, 34.862975, -85.428143, 34.866919],
+    icon: faMountain,
+  },
+  // Enterprise
+  {
+    trailName: 'Atlas',
+    displayName: 'Atlas',
+    recArea: 'Enterprise South',
+    rating: 'intermediate',
+    color: trailColor('intermediate', false),
+    distance: 5.4,
+    elevationGain: 888,
+    elevationLoss: 888,
+    elevationMin: 875,
+    elevationMax: 1024,
+    defaultBounds: [-85.124645, 35.074122, -85.107479, 35.082551],
+    icon: faMountain,
+  },
+  // Enterprise South Nature Park
+  {
+    trailName: 'ADA Accessible Trail',
+    displayName: 'ADA Accessible Trail',
+    recArea: 'Enterprise South',
+    rating: '',
+    color: trailColor('', false),
+    distance: 1.4,
+    elevationGain: 233,
+    elevationLoss: 221,
+    elevationMin: 956,
+    elevationMax: 1043,
+    defaultBounds: [-85.109711, 35.082551, -85.105247, 35.089012],
+    icon: faMountain,
+  },
+  {
+    trailName: 'Bicycles/Pedestrians',
+    displayName: 'Bicycles/Pedestrians',
+    recArea: 'Enterprise South',
+    rating: 'easy',
+    color: trailColor('easy', false),
+    distance: 6.6,
+    elevationGain: 1077,
+    elevationLoss: 982,
+    elevationMin: 823,
+    elevationMax: 1010,
+    defaultBounds: [-85.116405, 35.084517, -85.096493, 35.097018],
+    icon: faMountain,
+  },
+  {
+    trailName: 'Black Forest Bike Trail',
+    displayName: 'Black Forest Bike Trail',
+    recArea: 'Enterprise South',
+    rating: 'easy',
+    color: trailColor('easy', false),
+    distance: 3.2,
+    elevationGain: 588,
+    elevationLoss: 477,
+    elevationMin: 885,
+    elevationMax: 1024,
+    defaultBounds: [-85.119495, 35.078336, -85.108681, 35.084658],
+    icon: faMountain,
+  },
+  {
+    trailName: 'Boulder Point Trail',
+    displayName: 'Boulder Point Trail',
+    recArea: 'Enterprise South',
+    rating: '',
+    color: trailColor('', false),
+    distance: 1.3,
+    elevationGain: 221,
+    elevationLoss: 211,
+    elevationMin: 950,
+    elevationMax: 1023,
+    defaultBounds: [-85.113487, 35.090277, -85.104046, 35.094069],
+    icon: faMountain,
+  },
+  {
+    trailName: 'Deer Hollow Trail',
+    displayName: 'Deer Hollow Trail',
+    recArea: 'Enterprise South',
+    rating: '',
+    color: trailColor('', false),
+    distance: 0.6,
+    elevationGain: 96,
+    elevationLoss: 92,
+    elevationMin: 926,
+    elevationMax: 996,
+    defaultBounds: [-85.111771, 35.08522, -85.108337, 35.0914],
+    icon: faMountain,
+  },
+  {
+    trailName: 'Dogwood Ridge Trail',
+    displayName: 'Dogwood Ridge Trail',
+    recArea: 'Enterprise South',
+    rating: '',
+    color: trailColor('', false),
+    distance: 1.3,
+    elevationGain: 237,
+    elevationLoss: 276,
+    elevationMin: 879,
+    elevationMax: 990,
+    defaultBounds: [-85.115204, 35.089996, -85.106106, 35.096738],
+    icon: faMountain,
+  },
+  {
+    trailName: 'Hawks Ridge',
+    displayName: 'Hawks Ridge',
+    recArea: 'Enterprise South',
+    rating: '',
+    color: trailColor('', false),
+    distance: 3.8,
+    elevationGain: 789,
+    elevationLoss: 876,
+    elevationMin: 857,
+    elevationMax: 988,
+    defaultBounds: [-85.120869, 35.088029, -85.106106, 35.099125],
+    icon: faMountain,
+  },
+  {
+    trailName: 'Hidden Lake Overlook',
+    displayName: 'Hidden Lake Overlook',
+    recArea: 'Enterprise South',
+    rating: '',
+    color: trailColor('', false),
+    distance: 0.2,
+    elevationGain: 25,
+    elevationLoss: 61,
+    elevationMin: 953,
+    elevationMax: 1004,
+    defaultBounds: [-85.112457, 35.090136, -85.109711, 35.091541],
+    icon: faMountain,
+  },
+  {
+    trailName: 'Hidden Lake Trail',
+    displayName: 'Hidden Lake Trail',
+    recArea: 'Enterprise South',
+    rating: '',
+    color: trailColor('', false),
+    distance: 0.7,
+    elevationGain: 166,
+    elevationLoss: 125,
+    elevationMin: 901,
+    elevationMax: 1023,
+    defaultBounds: [-85.115376, 35.087748, -85.109196, 35.092243],
+    icon: faMountain,
+  },
+  {
+    trailName: 'Log Rhythm Bike Trail',
+    displayName: 'Log Rhythm Bike Trail',
+    recArea: 'Enterprise South',
+    rating: 'intermediate',
+    color: trailColor('intermediate', false),
+    distance: 3.6,
+    elevationGain: 719,
+    elevationLoss: 825,
+    elevationMin: 829,
+    elevationMax: 974,
+    defaultBounds: [-85.108509, 35.091541, -85.096321, 35.099827],
+    icon: faMountain,
+  },
+  {
+    trailName: 'Poe Run Overlook',
+    displayName: 'Poe Run Overlook',
+    recArea: 'Enterprise South',
+    rating: '',
+    color: trailColor('', false),
+    distance: 0.6,
+    elevationGain: 70,
+    elevationLoss: 125,
+    elevationMin: 891,
+    elevationMax: 987,
+    defaultBounds: [-85.115891, 35.084798, -85.111771, 35.086905],
+    icon: faMountain,
+  },
+  {
+    trailName: 'TNT Bike Trail',
+    displayName: 'TNT Bike Trail',
+    recArea: 'Enterprise South',
+    rating: 'easy',
+    color: trailColor('easy', false),
+    distance: 4.7,
+    elevationGain: 913,
+    elevationLoss: 1052,
+    elevationMin: 836,
+    elevationMax: 1053,
+    defaultBounds: [-85.109196, 35.083113, -85.099068, 35.092664],
+    icon: faMountain,
+  },
+  // Fletcher Park
+  {
+    trailName: 'Fletcher Park',
+    displayName: 'Fletcher Park',
+    recArea: 'Cleveland',
+    rating: '',
+    color: trailColor('', false),
+    distance: 1.8,
+    elevationGain: 126,
+    elevationLoss: 130,
+    elevationMin: 719,
+    elevationMax: 770,
+    defaultBounds: [-84.904575, 35.176615, -84.899597, 35.18377],
+    icon: faMountain,
+  },
+  // Greenway Farms
+  {
+    trailName: 'North Chickamauga Creek Greenway',
+    displayName: 'North Chickamauga Creek Greenway',
+    recArea: 'Greenway Farms',
+    rating: '',
+    color: trailColor('', true),
+    distance: 7.1,
+    elevationGain: 1180,
+    elevationLoss: 1150,
+    elevationMin: 635,
+    elevationMax: 775,
+    defaultBounds: [-85.228844, 35.110781, -85.214596, 35.153319],
+    icon: faRoute,
+  },
+  {
+    trailName: 'Wildflower Trail',
+    displayName: 'Wildflower Trail',
+    recArea: 'Greenway Farms',
+    rating: '',
+    color: trailColor('', true),
+    distance: 1.6,
+    elevationGain: 81,
+    elevationLoss: 91,
+    elevationMin: 639,
+    elevationMax: 675,
+    defaultBounds: [-85.227642, 35.12398, -85.218544, 35.128754],
+    icon: faRoute,
+  },
+  // Harrison Bay SP
+  {
+    trailName: 'BAY POINT LOOP TRAIL',
+    displayName: 'Bay Point Loop Trail',
+    recArea: 'Harrison Bay State Park',
+    rating: '',
+    color: trailColor('', false),
+    distance: 4.2,
+    elevationGain: 351,
+    elevationLoss: 356,
+    elevationMin: 683,
+    elevationMax: 716,
+    defaultBounds: [-85.123959, 35.155705, -85.114689, 35.168195],
+    icon: faMountain,
+  },
+  // Houston Valley ORV
+  {
+    trailName: 'Houston Valley ORV 181',
+    displayName: 'Houston Valley ORV 181',
+    recArea: 'Houston Valley',
+    rating: '',
+    color: trailColor('', false),
+    distance: 2.3,
+    elevationGain: 324,
+    elevationLoss: 279,
+    elevationMin: 948,
+    elevationMax: 1124,
+    defaultBounds: [-85.133915, 34.768691, -85.122585, 34.789841],
+    icon: faMountain,
+  },
+  {
+    trailName: 'Houston Valley ORV 181A',
+    displayName: 'Houston Valley ORV 181A',
+    recArea: 'Houston Valley',
+    rating: '',
+    color: trailColor('', false),
+    distance: 0.2,
+    elevationGain: 120,
+    elevationLoss: 2,
+    elevationMin: 975,
+    elevationMax: 1093,
+    defaultBounds: [-85.132713, 34.788995, -85.129452, 34.789982],
+    icon: faMountain,
+  },
+  {
+    trailName: 'Houston Valley ORV 181B',
+    displayName: 'Houston Valley ORV 181B',
+    recArea: 'Houston Valley',
+    rating: '',
+    color: trailColor('', false),
+    distance: 0.2,
+    elevationGain: 33,
+    elevationLoss: 39,
+    elevationMin: 965,
+    elevationMax: 996,
+    defaultBounds: [-85.130825, 34.785752, -85.130653, 34.788995],
+    icon: faMountain,
+  },
+  {
+    trailName: 'Houston Valley ORV 181C',
+    displayName: 'Houston Valley ORV 181C',
+    recArea: 'Houston Valley',
+    rating: '',
+    color: trailColor('', false),
+    distance: 1.0,
+    elevationGain: 273,
+    elevationLoss: 273,
+    elevationMin: 1029,
+    elevationMax: 1209,
+    defaultBounds: [-85.136318, 34.789982, -85.131683, 34.793506],
+    icon: faMountain,
+  },
+  {
+    trailName: 'Houston Valley ORV 181D',
+    displayName: 'Houston Valley ORV 181D',
+    recArea: 'Houston Valley',
+    rating: '',
+    color: trailColor('', false),
+    distance: 0.6,
+    elevationGain: 177,
+    elevationLoss: 124,
+    elevationMin: 1046,
+    elevationMax: 1167,
+    defaultBounds: [-85.136833, 34.791955, -85.132542, 34.794916],
+    icon: faMountain,
+  },
+  {
+    trailName: 'Houston Valley ORV 181E',
+    displayName: 'Houston Valley ORV 181E',
+    recArea: 'Houston Valley',
+    rating: '',
+    color: trailColor('', false),
+    distance: 1.2,
+    elevationGain: 430,
+    elevationLoss: 425,
+    elevationMin: 997,
+    elevationMax: 1197,
+    defaultBounds: [-85.137177, 34.78688, -85.132713, 34.792096],
+    icon: faMountain,
+  },
+  {
+    trailName: 'Houston Valley ORV 181F',
+    displayName: 'Houston Valley ORV 181F',
+    recArea: 'Houston Valley',
+    rating: '',
+    color: trailColor('', false),
+    distance: 2.0,
+    elevationGain: 601,
+    elevationLoss: 602,
+    elevationMin: 963,
+    elevationMax: 1205,
+    defaultBounds: [-85.139236, 34.773063, -85.132713, 34.787021],
+    icon: faMountain,
+  },
+  {
+    trailName: 'Houston Valley ORV 181G',
+    displayName: 'Houston Valley ORV 181G',
+    recArea: 'Houston Valley',
+    rating: '',
+    color: trailColor('', false),
+    distance: 0.2,
+    elevationGain: 114,
+    elevationLoss: 9,
+    elevationMin: 984,
+    elevationMax: 1097,
+    defaultBounds: [-85.133743, 34.784765, -85.131168, 34.786034],
+    icon: faMountain,
+  },
+  {
+    trailName: 'Houston Valley ORV 181H',
+    displayName: 'Houston Valley ORV 181H',
+    recArea: 'Houston Valley',
+    rating: '',
+    color: trailColor('', false),
+    distance: 0.4,
+    elevationGain: 130,
+    elevationLoss: 17,
+    elevationMin: 975,
+    elevationMax: 1088,
+    defaultBounds: [-85.13649, 34.77264, -85.133057, 34.774332],
+    icon: faMountain,
+  },
+  {
+    trailName: 'Houston Valley ORV 181J',
+    displayName: 'Houston Valley ORV 181J',
+    recArea: 'Houston Valley',
+    rating: '',
+    color: trailColor('', false),
+    distance: 2.1,
+    elevationGain: 518,
+    elevationLoss: 314,
+    elevationMin: 991,
+    elevationMax: 1196,
+    defaultBounds: [-85.143528, 34.76291, -85.133915, 34.772781],
+    icon: faMountain,
+  },
+  {
+    trailName: 'Houston Valley ORV 181K',
+    displayName: 'Houston Valley ORV 181K',
+    recArea: 'Houston Valley',
+    rating: '',
+    color: trailColor('', false),
+    distance: 0.5,
+    elevationGain: 116,
+    elevationLoss: 117,
+    elevationMin: 1079,
+    elevationMax: 1157,
+    defaultBounds: [-85.14061, 34.761217, -85.13546, 34.764038],
+    icon: faMountain,
+  },
+  {
+    trailName: 'Houston Valley ORV 181L',
+    displayName: 'Houston Valley ORV 181L',
+    recArea: 'Houston Valley',
+    rating: '',
+    color: trailColor('', false),
+    distance: 4.1,
+    elevationGain: 837,
+    elevationLoss: 992,
+    elevationMin: 1050,
+    elevationMax: 1411,
+    defaultBounds: [-85.142155, 34.760935, -85.11692, 34.767281],
+    icon: faMountain,
+  },
+  {
+    trailName: 'Houston Valley ORV 181M',
+    displayName: 'Houston Valley ORV 181M',
+    recArea: 'Houston Valley',
+    rating: '',
+    color: trailColor('', false),
+    distance: 1.0,
+    elevationGain: 342,
+    elevationLoss: 105,
+    elevationMin: 1113,
+    elevationMax: 1350,
+    defaultBounds: [-85.128937, 34.765166, -85.121212, 34.768691],
+    icon: faMountain,
+  },
+  {
+    trailName: 'Houston Valley ORV 181N',
+    displayName: 'Houston Valley ORV 181N',
+    recArea: 'Houston Valley',
+    rating: '',
+    color: trailColor('', false),
+    distance: 1.7,
+    elevationGain: 288,
+    elevationLoss: 280,
+    elevationMin: 938,
+    elevationMax: 1072,
+    defaultBounds: [-85.122929, 34.771653, -85.111256, 34.778562],
+    icon: faMountain,
+  },
+  {
+    trailName: 'Houston Valley ORV 181O',
+    displayName: 'Houston Valley ORV 181O',
+    recArea: 'Houston Valley',
+    rating: '',
+    color: trailColor('', false),
+    distance: 0.2,
+    elevationGain: 6,
+    elevationLoss: 123,
+    elevationMin: 1006,
+    elevationMax: 1128,
+    defaultBounds: [-85.12516, 34.773909, -85.122757, 34.774614],
+    icon: faMountain,
+  },
+  {
+    trailName: 'Houston Valley ORV 181P',
+    displayName: 'Houston Valley ORV 181P',
+    recArea: 'Houston Valley',
+    rating: '',
+    color: trailColor('', false),
+    distance: 2.0,
+    elevationGain: 514,
+    elevationLoss: 897,
+    elevationMin: 947,
+    elevationMax: 1367,
+    defaultBounds: [-85.126534, 34.771935, -85.111084, 34.784765],
+    icon: faMountain,
+  },
+  {
+    trailName: 'Houston Valley ORV 181Q',
+    displayName: 'Houston Valley ORV 181Q',
+    recArea: 'Houston Valley',
+    rating: '',
+    color: trailColor('', false),
+    distance: 0.7,
+    elevationGain: 417,
+    elevationLoss: 46,
+    elevationMin: 959,
+    elevationMax: 1331,
+    defaultBounds: [-85.131855, 34.770384, -85.126534, 34.776165],
+    icon: faMountain,
+  },
+  {
+    trailName: 'Houston Valley ORV 181R',
+    displayName: 'Houston Valley ORV 181R',
+    recArea: 'Houston Valley',
+    rating: '',
+    color: trailColor('', false),
+    distance: 0.4,
+    elevationGain: 57,
+    elevationLoss: 70,
+    elevationMin: 948,
+    elevationMax: 1006,
+    defaultBounds: [-85.133228, 34.776165, -85.12825, 34.777857],
+    icon: faMountain,
+  },
+  {
+    trailName: 'Houston Valley ORV 181S',
+    displayName: 'Houston Valley ORV 181S',
+    recArea: 'Houston Valley',
+    rating: '',
+    color: trailColor('', false),
+    distance: 0.7,
+    elevationGain: 129,
+    elevationLoss: 129,
+    elevationMin: 946,
+    elevationMax: 1024,
+    defaultBounds: [-85.128422, 34.777857, -85.126019, 34.781382],
+    icon: faMountain,
+  },
+  {
+    trailName: 'Houston Valley ORV 181T',
+    displayName: 'Houston Valley ORV 181T',
+    recArea: 'Houston Valley',
+    rating: '',
+    color: trailColor('', false),
+    distance: 1.0,
+    elevationGain: 106,
+    elevationLoss: 100,
+    elevationMin: 942,
+    elevationMax: 985,
+    defaultBounds: [-85.132713, 34.777152, -85.127907, 34.788149],
+    icon: faMountain,
+  },
+  {
+    trailName: 'Houston Valley ORV 181U',
+    displayName: 'Houston Valley ORV 181U',
+    recArea: 'Houston Valley',
+    rating: '',
+    color: trailColor('', false),
+    distance: 0.0,
+    elevationGain: 8,
+    elevationLoss: 0,
+    elevationMin: 974,
+    elevationMax: 982,
+    defaultBounds: [-85.130825, 34.78547, -85.13031, 34.785752],
+    icon: faMountain,
+  },
+  // Longbranch
+  {
+    trailName: 'Long Branch Trail',
+    displayName: 'Long Branch Trail',
+    recArea: 'Durham',
+    rating: 'intermediate',
+    color: trailColor('intermediate', false),
+    distance: 6.9,
+    elevationGain: 1478,
+    elevationLoss: 1187,
+    elevationMin: 1431,
+    elevationMax: 1966,
+    defaultBounds: [-85.413551, 34.855228, -85.391064, 34.897758],
+    icon: faMountain,
+  },
+  // Lookout Mountain
+  {
+    trailName: 'Cravens Terrace Road',
+    displayName: 'Cravens Terrace Road',
+    recArea: 'Lookout Mountain',
+    rating: 'easy',
+    color: trailColor('easy', false),
+    distance: 1.2,
+    elevationGain: 133,
+    elevationLoss: 595,
+    elevationMin: 985,
+    elevationMax: 1447,
+    defaultBounds: [-85.344372, 35.014252, -85.335274, 35.016923],
+    icon: faMountain,
+  },
+  {
+    trailName: 'Guild Trail',
+    displayName: 'Guild Trail',
+    recArea: 'Lookout Mountain',
+    rating: 'easy',
+    color: trailColor('easy', false),
+    distance: 3.3,
+    elevationGain: 884,
+    elevationLoss: 435,
+    elevationMin: 858,
+    elevationMax: 1307,
+    defaultBounds: [-85.3545, 35.004128, -85.331669, 35.020719],
+    icon: faMountain,
+  },
+  {
+    trailName: 'Hardy Trail',
+    displayName: 'Hardy Trail',
+    recArea: 'Lookout Mountain',
+    rating: 'easy',
+    color: trailColor('easy', false),
+    distance: 1.5,
+    elevationGain: 406,
+    elevationLoss: 128,
+    elevationMin: 1451,
+    elevationMax: 1729,
+    defaultBounds: [-85.3442, 34.995973, -85.340252, 35.015939],
+    icon: faMountain,
+  },
+  {
+    trailName: 'Jackson Gap Trail',
+    displayName: 'Jackson Gap Trail',
+    recArea: 'Lookout Mountain',
+    rating: 'advanced',
+    color: trailColor('advanced', false),
+    distance: 3.0,
+    elevationGain: 1391,
+    elevationLoss: 879,
+    elevationMin: 1293,
+    elevationMax: 1952,
+    defaultBounds: [-85.384712, 34.957573, -85.376816, 34.970094],
+    icon: faMountain,
+  },
+  {
+    trailName: 'Lower Truck Trail',
+    displayName: 'Lower Truck Trail',
+    recArea: 'Lookout Mountain',
+    rating: 'easy',
+    color: trailColor('easy', false),
+    distance: 2.3,
+    elevationGain: 293,
+    elevationLoss: 302,
+    elevationMin: 646,
+    elevationMax: 697,
+    defaultBounds: [-85.388145, 34.96939, -85.372181, 34.997379],
+    icon: faMountain,
+  },
+  {
+    trailName: 'Southend Trail',
+    displayName: 'Southend Trail',
+    recArea: 'Lookout Mountain',
+    rating: 'easy',
+    color: trailColor('easy', false),
+    distance: 0.6,
+    elevationGain: 229,
+    elevationLoss: 274,
+    elevationMin: 640,
+    elevationMax: 719,
+    defaultBounds: [-85.394154, 34.962919, -85.388145, 34.96939],
+    icon: faMountain,
+  },
+  {
+    trailName: 'Upper Truck Trail',
+    displayName: 'Upper Truck Trail',
+    recArea: 'Lookout Mountain',
+    rating: 'easy',
+    color: trailColor('easy', false),
+    distance: 4.6,
+    elevationGain: 1079,
+    elevationLoss: 924,
+    elevationMin: 1220,
+    elevationMax: 1449,
+    defaultBounds: [-85.384197, 34.965592, -85.344372, 35.015939],
+    icon: faMountain,
+  },
+  // Lula Lake
+  {
+    trailName: 'Bluff Road',
+    displayName: 'Bluff Road',
+    recArea: 'Lula Lake',
+    rating: 'easy',
+    color: trailColor('easy', false),
+    distance: 1.8,
+    elevationGain: 533,
+    elevationLoss: 134,
+    elevationMin: 1281,
+    elevationMax: 1732,
+    defaultBounds: [-85.378189, 34.912962, -85.369091, 34.93309],
+    icon: faMountain,
+  },
+  {
+    trailName: 'F4',
+    displayName: 'F4',
+    recArea: 'Lula Lake',
+    rating: 'intermediate',
+    color: trailColor('intermediate', false),
+    distance: 1.1,
+    elevationGain: 331,
+    elevationLoss: 291,
+    elevationMin: 1328,
+    elevationMax: 1495,
+    defaultBounds: [-85.386944, 34.902404, -85.380936, 34.915215],
+    icon: faMountain,
+  },
+  {
+    trailName: 'South Creek',
+    displayName: 'South Creek',
+    recArea: 'Lula Lake',
+    rating: 'intermediate',
+    color: trailColor('intermediate', false),
+    distance: 0.8,
+    elevationGain: 207,
+    elevationLoss: 83,
+    elevationMin: 1321,
+    elevationMax: 1456,
+    defaultBounds: [-85.383854, 34.914652, -85.380077, 34.920282],
+    icon: faMountain,
+  },
+  // Lula Lake Land Trust
+  {
+    trailName: 'Access Road / Old RR Grade',
+    displayName: 'Access Road / Old RR Grade',
+    recArea: 'Lula Lake',
+    rating: 'easy',
+    color: trailColor('easy', false),
+    distance: 2.4,
+    elevationGain: 579,
+    elevationLoss: 779,
+    elevationMin: 1125,
+    elevationMax: 1510,
+    defaultBounds: [-85.386257, 34.91803, -85.370121, 34.934215],
+    icon: faMountain,
+  },
+  {
+    trailName: 'Jedi Trail',
+    displayName: 'Jedi Trail',
+    recArea: 'Lula Lake',
+    rating: 'intermediate',
+    color: trailColor('intermediate', false),
+    distance: 0.9,
+    elevationGain: 72,
+    elevationLoss: 174,
+    elevationMin: 1460,
+    elevationMax: 1606,
+    defaultBounds: [-85.376816, 34.920423, -85.369778, 34.931401],
+    icon: faMountain,
+  },
+  {
+    trailName: 'Long Branch Trail Horse Ford',
+    displayName: 'Long Branch Trail Horse Ford',
+    recArea: 'Lula Lake',
+    rating: 'intermediate',
+    color: trailColor('intermediate', false),
+    distance: 0.1,
+    elevationGain: 4,
+    elevationLoss: 12,
+    elevationMin: 1633,
+    elevationMax: 1645,
+    defaultBounds: [-85.395699, 34.875651, -85.395527, 34.876637],
+    icon: faMountain,
+  },
+  {
+    trailName: 'Middle Trail',
+    displayName: 'Middle Trail',
+    recArea: 'Lula Lake',
+    rating: 'intermediate',
+    color: trailColor('intermediate', false),
+    distance: 2.8,
+    elevationGain: 582,
+    elevationLoss: 635,
+    elevationMin: 1329,
+    elevationMax: 1497,
+    defaultBounds: [-85.383511, 34.912962, -85.369434, 34.931401],
+    icon: faMountain,
+  },
+  {
+    trailName: 'Ovenbird Trail',
+    displayName: 'Ovenbird Trail',
+    recArea: 'Lula Lake',
+    rating: 'intermediate',
+    color: trailColor('intermediate', false),
+    distance: 0.3,
+    elevationGain: 7,
+    elevationLoss: 101,
+    elevationMin: 1493,
+    elevationMax: 1587,
+    defaultBounds: [-85.379562, 34.918311, -85.376816, 34.920423],
+    icon: faMountain,
+  },
+  {
+    trailName: 'Parking Loop',
+    displayName: 'Parking Loop',
+    recArea: 'Lula Lake',
+    rating: 'easy',
+    color: trailColor('easy', false),
+    distance: 0.2,
+    elevationGain: 6,
+    elevationLoss: 52,
+    elevationMin: 1313,
+    elevationMax: 1359,
+    defaultBounds: [-85.381451, 34.921971, -85.379219, 34.92366],
+    icon: faMountain,
+  },
+  {
+    trailName: 'Spur and Connector Trails',
+    displayName: 'Spur and Connector Trails',
+    recArea: 'Lula Lake',
+    rating: 'intermediate',
+    color: trailColor('intermediate', false),
+    distance: 1.7,
+    elevationGain: 524,
+    elevationLoss: 661,
+    elevationMin: 1134,
+    elevationMax: 1600,
+    defaultBounds: [-85.381107, 34.920001, -85.371151, 34.93309],
+    icon: faMountain,
+  },
+  {
+    trailName: 'Trail A',
+    displayName: 'Trail A',
+    recArea: 'Lula Lake',
+    rating: 'intermediate',
+    color: trailColor('intermediate', false),
+    distance: 0.5,
+    elevationGain: 168,
+    elevationLoss: 142,
+    elevationMin: 1282,
+    elevationMax: 1347,
+    defaultBounds: [-85.377846, 34.926616, -85.373554, 34.930979],
+    icon: faMountain,
+  },
+  {
+    trailName: 'Trail C',
+    displayName: 'Trail C',
+    recArea: 'Lula Lake',
+    rating: 'intermediate',
+    color: trailColor('intermediate', false),
+    distance: 0.4,
+    elevationGain: 44,
+    elevationLoss: 122,
+    elevationMin: 1319,
+    elevationMax: 1405,
+    defaultBounds: [-85.384026, 34.919578, -85.381279, 34.92169],
+    icon: faMountain,
+  },
+  {
+    trailName: 'Turkey Trail',
+    displayName: 'Turkey Trail',
+    recArea: 'Lula Lake',
+    rating: 'intermediate',
+    color: trailColor('intermediate', false),
+    distance: 0.9,
+    elevationGain: 43,
+    elevationLoss: 332,
+    elevationMin: 1393,
+    elevationMax: 1713,
+    defaultBounds: [-85.377502, 34.916482, -85.375099, 34.927179],
+    icon: faMountain,
+  },
+  // Lula Lake Land Trust 5 Points
+  {
+    trailName: 'Atlatl',
+    displayName: 'Atlatl',
+    recArea: '5 Points',
+    rating: 'intermediate',
+    color: trailColor('intermediate', false),
+    distance: 1.3,
+    elevationGain: 320,
+    elevationLoss: 223,
+    elevationMin: 1648,
+    elevationMax: 1784,
+    defaultBounds: [-85.46299, 34.82085, -85.455608, 34.825641],
+    icon: faMountain,
+  },
+  {
+    trailName: 'Back End',
+    displayName: 'Back End',
+    recArea: '5 Points',
+    rating: 'easy',
+    color: trailColor('easy', false),
+    distance: 1.2,
+    elevationGain: 305,
+    elevationLoss: 223,
+    elevationMin: 2001,
+    elevationMax: 2251,
+    defaultBounds: [-85.440845, 34.846212, -85.429344, 34.85072],
+    icon: faMountain,
+  },
+  {
+    trailName: 'Bankhead',
+    displayName: 'Bankhead',
+    recArea: '5 Points',
+    rating: 'easy',
+    color: trailColor('easy', false),
+    distance: 0.6,
+    elevationGain: 89,
+    elevationLoss: 119,
+    elevationMin: 1911,
+    elevationMax: 1997,
+    defaultBounds: [-85.429173, 34.85072, -85.424023, 34.850861],
+    icon: faMountain,
+  },
+  {
+    trailName: 'Barkeater',
+    displayName: 'Barkeater',
+    recArea: '5 Points',
+    rating: 'intermediate',
+    color: trailColor('intermediate', false),
+    distance: 3.1,
+    elevationGain: 592,
+    elevationLoss: 685,
+    elevationMin: 1635,
+    elevationMax: 1885,
+    defaultBounds: [-85.455265, 34.824937, -85.441017, 34.835364],
+    icon: faMountain,
+  },
+  {
+    trailName: "Bella's Run",
+    displayName: "Bella's Run",
+    recArea: '5 Points',
+    rating: 'intermediate',
+    color: trailColor('intermediate', false),
+    distance: 1.2,
+    elevationGain: 325,
+    elevationLoss: 259,
+    elevationMin: 2093,
+    elevationMax: 2190,
+    defaultBounds: [-85.438786, 34.845085, -85.430374, 34.849875],
+    icon: faMountain,
+  },
+  {
+    trailName: 'Caprock',
+    displayName: 'Caprock',
+    recArea: '5 Points',
+    rating: 'advanced',
+    color: trailColor('advanced', false),
+    distance: 1.1,
+    elevationGain: 234,
+    elevationLoss: 419,
+    elevationMin: 1895,
+    elevationMax: 2091,
+    defaultBounds: [-85.44445, 34.840295, -85.439987, 34.846494],
+    icon: faMountain,
+  },
+  {
+    trailName: 'Caprock Cut Through',
+    displayName: 'Caprock Cut Through',
+    recArea: '5 Points',
+    rating: 'easy',
+    color: trailColor('easy', false),
+    distance: 0.2,
+    elevationGain: 29,
+    elevationLoss: 36,
+    elevationMin: 1951,
+    elevationMax: 1976,
+    defaultBounds: [-85.441189, 34.84269, -85.439472, 34.844099],
+    icon: faMountain,
+  },
+  {
+    trailName: 'Cross Cut',
+    displayName: 'Cross Cut',
+    recArea: '5 Points',
+    rating: 'intermediate',
+    color: trailColor('intermediate', false),
+    distance: 1.6,
+    elevationGain: 463,
+    elevationLoss: 577,
+    elevationMin: 1982,
+    elevationMax: 2249,
+    defaultBounds: [-85.440845, 34.846494, -85.429344, 34.850861],
+    icon: faMountain,
+  },
+  {
+    trailName: 'Foot Wall',
+    displayName: 'Foot Wall',
+    recArea: '5 Points',
+    rating: 'intermediate',
+    color: trailColor('intermediate', false),
+    distance: 0.9,
+    elevationGain: 363,
+    elevationLoss: 368,
+    elevationMin: 1954,
+    elevationMax: 2032,
+    defaultBounds: [-85.430031, 34.851002, -85.428143, 34.861848],
+    icon: faMountain,
+  },
+  {
+    trailName: 'Fuggett Lift',
+    displayName: 'Fuggett Lift',
+    recArea: '5 Points',
+    rating: 'intermediate',
+    color: trailColor('intermediate', false),
+    distance: 0.4,
+    elevationGain: 195,
+    elevationLoss: 74,
+    elevationMin: 1962,
+    elevationMax: 2103,
+    defaultBounds: [-85.427456, 34.860158, -85.425224, 34.86213],
+    icon: faMountain,
+  },
+  {
+    trailName: 'Hanging Wall',
+    displayName: 'Hanging Wall',
+    recArea: '5 Points',
+    rating: 'intermediate',
+    color: trailColor('intermediate', false),
+    distance: 0.4,
+    elevationGain: 287,
+    elevationLoss: 179,
+    elevationMin: 2007,
+    elevationMax: 2123,
+    defaultBounds: [-85.429001, 34.857059, -85.427456, 34.861708],
+    icon: faMountain,
+  },
+  {
+    trailName: 'Kettle Bottom',
+    displayName: 'Kettle Bottom',
+    recArea: '5 Points',
+    rating: 'intermediate',
+    color: trailColor('intermediate', false),
+    distance: 4.3,
+    elevationGain: 1129,
+    elevationLoss: 1120,
+    elevationMin: 1886,
+    elevationMax: 2000,
+    defaultBounds: [-85.432606, 34.85072, -85.428829, 34.86213],
+    icon: faMountain,
+  },
+  {
+    trailName: 'Kinder Garden',
+    displayName: 'Kinder Garden',
+    recArea: '5 Points',
+    rating: 'advanced',
+    color: trailColor('advanced', false),
+    distance: 0.5,
+    elevationGain: 57,
+    elevationLoss: 160,
+    elevationMin: 1765,
+    elevationMax: 1873,
+    defaultBounds: [-85.450802, 34.830009, -85.447712, 34.831137],
+    icon: faMountain,
+  },
+  {
+    trailName: "Mick's Trail",
+    displayName: "Mick's Trail",
+    recArea: '5 Points',
+    rating: 'easy',
+    color: trailColor('easy', false),
+    distance: 1.6,
+    elevationGain: 636,
+    elevationLoss: 681,
+    elevationMin: 2010,
+    elevationMax: 2128,
+    defaultBounds: [-85.44445, 34.843958, -85.429344, 34.850439],
+    icon: faMountain,
+  },
+  {
+    trailName: 'North Hogsback',
+    displayName: 'North Hogsback',
+    recArea: '5 Points',
+    rating: 'intermediate',
+    color: trailColor('intermediate', false),
+    distance: 1.7,
+    elevationGain: 252,
+    elevationLoss: 241,
+    elevationMin: 1905,
+    elevationMax: 1976,
+    defaultBounds: [-85.430546, 34.84593, -85.424709, 34.851143],
+    icon: faMountain,
+  },
+  {
+    trailName: 'Peace Can',
+    displayName: 'Peace Can',
+    recArea: '5 Points',
+    rating: 'easy',
+    color: trailColor('easy', false),
+    distance: 1.1,
+    elevationGain: 177,
+    elevationLoss: 181,
+    elevationMin: 1895,
+    elevationMax: 1952,
+    defaultBounds: [-85.426083, 34.851284, -85.422649, 34.860581],
+    icon: faMountain,
+  },
+  {
+    trailName: 'Pony Boy Spur',
+    displayName: 'Pony Boy Spur',
+    recArea: '5 Points',
+    rating: 'intermediate',
+    color: trailColor('intermediate', false),
+    distance: 0.2,
+    elevationGain: 94,
+    elevationLoss: 0,
+    elevationMin: 1776,
+    elevationMax: 1870,
+    defaultBounds: [-85.449257, 34.828741, -85.447712, 34.83015],
+    icon: faMountain,
+  },
+  {
+    trailName: 'Prince Albert',
+    displayName: 'Prince Albert',
+    recArea: '5 Points',
+    rating: 'easy',
+    color: trailColor('easy', false),
+    distance: 1.1,
+    elevationGain: 174,
+    elevationLoss: 246,
+    elevationMin: 1768,
+    elevationMax: 1897,
+    defaultBounds: [-85.480156, 34.818736, -85.46711, 34.821414],
+    icon: faMountain,
+  },
+  {
+    trailName: 'Shale Flats',
+    displayName: 'Shale Flats',
+    recArea: '5 Points',
+    rating: 'easy',
+    color: trailColor('easy', false),
+    distance: 1.3,
+    elevationGain: 191,
+    elevationLoss: 212,
+    elevationMin: 1919,
+    elevationMax: 1976,
+    defaultBounds: [-85.440159, 34.840154, -85.430202, 34.847198],
+    icon: faMountain,
+  },
+  {
+    trailName: 'Slickenside',
+    displayName: 'Slickenside',
+    recArea: '5 Points',
+    rating: 'intermediate',
+    color: trailColor('intermediate', false),
+    distance: 1.2,
+    elevationGain: 331,
+    elevationLoss: 380,
+    elevationMin: 1974,
+    elevationMax: 2034,
+    defaultBounds: [-85.429344, 34.85072, -85.425911, 34.861708],
+    icon: faMountain,
+  },
+  {
+    trailName: 'South Hogsback',
+    displayName: 'South Hogsback',
+    recArea: '5 Points',
+    rating: 'advanced',
+    color: trailColor('advanced', false),
+    distance: 0.3,
+    elevationGain: 35,
+    elevationLoss: 81,
+    elevationMin: 1881,
+    elevationMax: 1927,
+    defaultBounds: [-85.440674, 34.838886, -85.438614, 34.840154],
+    icon: faMountain,
+  },
+  {
+    trailName: 'Tailings Run',
+    displayName: 'Tailings Run',
+    recArea: '5 Points',
+    rating: 'intermediate',
+    color: trailColor('intermediate', false),
+    distance: 1.5,
+    elevationGain: 270,
+    elevationLoss: 311,
+    elevationMin: 1851,
+    elevationMax: 1962,
+    defaultBounds: [-85.425224, 34.851143, -85.420933, 34.861567],
+    icon: faMountain,
+  },
+  {
+    trailName: 'Torino Hill',
+    displayName: 'Torino Hill',
+    recArea: '5 Points',
+    rating: 'intermediate',
+    color: trailColor('intermediate', false),
+    distance: 0.5,
+    elevationGain: 49,
+    elevationLoss: 158,
+    elevationMin: 2116,
+    elevationMax: 2225,
+    defaultBounds: [-85.442562, 34.844944, -85.438786, 34.847621],
+    icon: faMountain,
+  },
+  {
+    trailName: 'Windy Shot',
+    displayName: 'Windy Shot',
+    recArea: '5 Points',
+    rating: 'easy',
+    color: trailColor('easy', false),
+    distance: 0.8,
+    elevationGain: 125,
+    elevationLoss: 207,
+    elevationMin: 1999,
+    elevationMax: 2146,
+    defaultBounds: [-85.429344, 34.85072, -85.427284, 34.861989],
+    icon: faMountain,
+  },
+  // Moonshine
+  {
+    trailName: 'Bathtub Gin',
+    displayName: 'Bathtub Gin',
+    recArea: 'Lula Lake',
+    rating: 'advanced',
+    color: trailColor('advanced', false),
+    distance: 0.5,
+    elevationGain: 146,
+    elevationLoss: 139,
+    elevationMin: 1482,
+    elevationMax: 1552,
+    defaultBounds: [-85.388832, 34.930697, -85.384026, 34.93309],
+    icon: faMountain,
+  },
+  {
+    trailName: 'CCT',
+    displayName: 'CCT',
+    recArea: 'Lula Lake',
+    rating: 'easy',
+    color: trailColor('easy', false),
+    distance: 4.2,
+    elevationGain: 1056,
+    elevationLoss: 590,
+    elevationMin: 1406,
+    elevationMax: 1982,
+    defaultBounds: [-85.388145, 34.920704, -85.377331, 34.957292],
+    icon: faMountain,
+  },
+  {
+    trailName: 'Firewater',
+    displayName: 'Firewater',
+    recArea: 'Lula Lake',
+    rating: 'intermediate',
+    color: trailColor('intermediate', false),
+    distance: 1.9,
+    elevationGain: 370,
+    elevationLoss: 437,
+    elevationMin: 1413,
+    elevationMax: 1573,
+    defaultBounds: [-85.389519, 34.926475, -85.383854, 34.933512],
+    icon: faMountain,
+  },
+  {
+    trailName: 'White Lightning',
+    displayName: 'White Lightning',
+    recArea: 'Lula Lake',
+    rating: 'intermediate',
+    color: trailColor('intermediate', false),
+    distance: 1.4,
+    elevationGain: 163,
+    elevationLoss: 418,
+    elevationMin: 1395,
+    elevationMax: 1693,
+    defaultBounds: [-85.382996, 34.924786, -85.377846, 34.939],
+    icon: faMountain,
+  },
+  // Other
+  {
+    trailName: 'Box Turtle',
+    displayName: 'Box Turtle',
+    recArea: 'White Oak Mountain',
+    rating: 'intermediate',
+    color: trailColor('intermediate', false),
+    distance: 4.9,
+    elevationGain: 950,
+    elevationLoss: 1156,
+    elevationMin: 816,
+    elevationMax: 1086,
+    defaultBounds: [-85.041218, 35.031402, -85.033836, 35.042787],
+    icon: faMountain,
+  },
+  {
+    trailName: 'Mindbender',
+    displayName: 'Mindbender',
+    recArea: 'White Oak Mountain',
+    rating: 'intermediate',
+    color: trailColor('intermediate', false),
+    distance: 0.6,
+    elevationGain: 55,
+    elevationLoss: 246,
+    elevationMin: 892,
+    elevationMax: 1083,
+    defaultBounds: [-85.042419, 35.042927, -85.038986, 35.045457],
+    icon: faMountain,
+  },
+  {
+    trailName: 'Nutty Buddy (lower)',
+    displayName: 'Nutty Buddy (lower)',
+    recArea: 'White Oak Mountain',
+    rating: 'advanced',
+    color: trailColor('advanced', false),
+    distance: 0.5,
+    elevationGain: 39,
+    elevationLoss: 216,
+    elevationMin: 802,
+    elevationMax: 979,
+    defaultBounds: [-85.059414, 35.051359, -85.057182, 35.055013],
+    icon: faMountain,
+  },
+  // Raccoon Mountain
+  {
+    trailName: 'Boat Ramp Connector',
+    displayName: 'Boat Ramp Connector',
+    recArea: 'Raccoon Mountain',
+    rating: 'intermediate',
+    color: trailColor('intermediate', false),
+    distance: 0.3,
+    elevationGain: 48,
+    elevationLoss: 50,
+    elevationMin: 655,
+    elevationMax: 680,
+    defaultBounds: [-85.419216, 35.057261, -85.415268, 35.059791],
+    icon: faMountain,
+  },
+  {
+    trailName: 'Chunky',
+    displayName: 'Chunky',
+    recArea: 'Raccoon Mountain',
+    rating: 'advanced',
+    color: trailColor('advanced', false),
+    distance: 1.5,
+    elevationGain: 303,
+    elevationLoss: 417,
+    elevationMin: 1718,
+    elevationMax: 1842,
+    defaultBounds: [-85.384026, 35.048127, -85.380936, 35.051922],
+    icon: faMountain,
+  },
+  {
+    trailName: 'East Rim',
+    displayName: 'East Rim',
+    recArea: 'Raccoon Mountain',
+    rating: 'intermediate',
+    color: trailColor('intermediate', false),
+    distance: 1.2,
+    elevationGain: 325,
+    elevationLoss: 232,
+    elevationMin: 1689,
+    elevationMax: 1865,
+    defaultBounds: [-85.385399, 35.042506, -85.378704, 35.051922],
+    icon: faMountain,
+  },
+  {
+    trailName: 'Electric Avenue',
+    displayName: 'Electric Avenue',
+    recArea: 'Raccoon Mountain',
+    rating: 'easy',
+    color: trailColor('easy', false),
+    distance: 3.2,
+    elevationGain: 742,
+    elevationLoss: 500,
+    elevationMin: 1535,
+    elevationMax: 1786,
+    defaultBounds: [-85.419388, 35.038851, -85.385399, 35.047987],
+    icon: faMountain,
+  },
+  {
+    trailName: 'Grindstone Ridge',
+    displayName: 'Grindstone Ridge',
+    recArea: 'Raccoon Mountain',
+    rating: 'intermediate',
+    color: trailColor('intermediate', false),
+    distance: 1.6,
+    elevationGain: 392,
+    elevationLoss: 375,
+    elevationMin: 1641,
+    elevationMax: 1859,
+    defaultBounds: [-85.405483, 35.056278, -85.388832, 35.057683],
+    icon: faMountain,
+  },
+  {
+    trailName: 'High Voltage',
+    displayName: 'High Voltage',
+    recArea: 'Raccoon Mountain',
+    rating: 'intermediate',
+    color: trailColor('intermediate', false),
+    distance: 3.3,
+    elevationGain: 587,
+    elevationLoss: 1441,
+    elevationMin: 640,
+    elevationMax: 1510,
+    defaultBounds: [-85.424366, 35.04096, -85.415268, 35.057261],
+    icon: faMountain,
+  },
+  {
+    trailName: 'Laurel Point',
+    displayName: 'Laurel Point',
+    recArea: 'Raccoon Mountain',
+    rating: 'intermediate',
+    color: trailColor('intermediate', false),
+    distance: 2.6,
+    elevationGain: 523,
+    elevationLoss: 467,
+    elevationMin: 1510,
+    elevationMax: 1632,
+    defaultBounds: [-85.419559, 35.040538, -85.403938, 35.045738],
+    icon: faMountain,
+  },
+  {
+    trailName: 'Live Wire 1',
+    displayName: 'Live Wire 1',
+    recArea: 'Raccoon Mountain',
+    rating: 'intermediate',
+    color: trailColor('intermediate', false),
+    distance: 1.8,
+    elevationGain: 965,
+    elevationLoss: 382,
+    elevationMin: 1252,
+    elevationMax: 1835,
+    defaultBounds: [-85.385227, 35.051781, -85.378361, 35.06246],
+    icon: faMountain,
+  },
+  {
+    trailName: 'Live Wire 2',
+    displayName: 'Live Wire 2',
+    recArea: 'Raccoon Mountain',
+    rating: 'advanced',
+    color: trailColor('advanced', false),
+    distance: 2.9,
+    elevationGain: 1056,
+    elevationLoss: 456,
+    elevationMin: 657,
+    elevationMax: 1259,
+    defaultBounds: [-85.419216, 35.054873, -85.385227, 35.063444],
+    icon: faMountain,
+  },
+  {
+    trailName: 'Lower Chunky',
+    displayName: 'Lower Chunky',
+    recArea: 'Raccoon Mountain',
+    rating: 'advanced',
+    color: trailColor('advanced', false),
+    distance: 0.6,
+    elevationGain: 274,
+    elevationLoss: 130,
+    elevationMin: 1624,
+    elevationMax: 1814,
+    defaultBounds: [-85.386257, 35.052062, -85.380936, 35.056699],
+    icon: faMountain,
+  },
+  {
+    trailName: 'Megawatt',
+    displayName: 'Megawatt',
+    recArea: 'Raccoon Mountain',
+    rating: 'advanced',
+    color: trailColor('advanced', false),
+    distance: 2.5,
+    elevationGain: 729,
+    elevationLoss: 933,
+    elevationMin: 1429,
+    elevationMax: 1689,
+    defaultBounds: [-85.413208, 35.048268, -85.405483, 35.056559],
+    icon: faMountain,
+  },
+  {
+    trailName: 'Small Intestine Loop',
+    displayName: 'Small Intestine Loop',
+    recArea: 'Raccoon Mountain',
+    rating: 'easy',
+    color: trailColor('easy', false),
+    distance: 2.8,
+    elevationGain: 562,
+    elevationLoss: 577,
+    elevationMin: 1528,
+    elevationMax: 1663,
+    defaultBounds: [-85.391579, 35.05684, -85.384541, 35.061336],
+    icon: faMountain,
+  },
+  {
+    trailName: 'Split Rock',
+    displayName: 'Split Rock',
+    recArea: 'Raccoon Mountain',
+    rating: 'intermediate',
+    color: trailColor('intermediate', false),
+    distance: 0.4,
+    elevationGain: 80,
+    elevationLoss: 82,
+    elevationMin: 1498,
+    elevationMax: 1550,
+    defaultBounds: [-85.415611, 35.042225, -85.413895, 35.044052],
+    icon: faMountain,
+  },
+  {
+    trailName: 'Switchyard Connector',
+    displayName: 'Switchyard Connector',
+    recArea: 'Raccoon Mountain',
+    rating: 'intermediate',
+    color: trailColor('intermediate', false),
+    distance: 1.6,
+    elevationGain: 114,
+    elevationLoss: 286,
+    elevationMin: 1661,
+    elevationMax: 1840,
+    defaultBounds: [-85.38969, 35.051641, -85.379906, 35.058526],
+    icon: faMountain,
+  },
+  {
+    trailName: 'Tennessee River Gorge',
+    displayName: 'Tennessee River Gorge',
+    recArea: 'Raccoon Mountain',
+    rating: 'intermediate',
+    color: trailColor('intermediate', false),
+    distance: 1.4,
+    elevationGain: 348,
+    elevationLoss: 306,
+    elevationMin: 1474,
+    elevationMax: 1602,
+    defaultBounds: [-85.419903, 35.044473, -85.408058, 35.048549],
+    icon: faMountain,
+  },
+  // Ringgold Greenway
+  {
+    trailName: 'Chief Richard Taylor Natural trail',
+    displayName: 'Chief Richard Taylor Natural Trail',
+    recArea: 'Ringgold',
+    rating: '',
+    color: trailColor('', true),
+    distance: 0.6,
+    elevationGain: 6,
+    elevationLoss: 9,
+    elevationMin: 736,
+    elevationMax: 740,
+    defaultBounds: [-85.113659, 34.908317, -85.104218, 34.911133],
+    icon: faRoute,
+  },
+  {
+    trailName: 'Chief Richard Taylor Nature Trail',
+    displayName: 'Chief Richard Taylor Nature Trail',
+    recArea: 'Ringgold',
+    rating: '',
+    color: trailColor('', true),
+    distance: 1.1,
+    elevationGain: 67,
+    elevationLoss: 46,
+    elevationMin: 737,
+    elevationMax: 760,
+    defaultBounds: [-85.118122, 34.911133, -85.110912, 34.914511],
+    icon: faRoute,
+  },
+  // South Chattanooga communities
+  {
+    trailName: 'Crabtree Farm Trail Connection',
+    displayName: 'Crabtree Farm Trail Connection',
+    recArea: 'Lookout Mountain',
+    rating: '',
+    color: trailColor('', true),
+    distance: 1.3,
+    elevationGain: 105,
+    elevationLoss: 103,
+    elevationMin: 638,
+    elevationMax: 668,
+    defaultBounds: [-85.309696, 35.008487, -85.299225, 35.012283],
+    icon: faRoute,
+  },
+  {
+    trailName: 'Crabtree Farms / East Lake Connector',
+    displayName: 'Crabtree Farms / East Lake Connector',
+    recArea: 'Lookout Mountain',
+    rating: '',
+    color: trailColor('', true),
+    distance: 1.6,
+    elevationGain: 85,
+    elevationLoss: 48,
+    elevationMin: 651,
+    elevationMax: 702,
+    defaultBounds: [-85.300083, 35.001457, -85.278625, 35.009471],
+    icon: faRoute,
+  },
+  {
+    trailName: 'Guild-Hardy Connector',
+    displayName: 'Guild-Hardy Connector',
+    recArea: 'Lookout Mountain',
+    rating: 'easy',
+    color: trailColor('easy', false),
+    distance: 0.8,
+    elevationGain: 40,
+    elevationLoss: 356,
+    elevationMin: 677,
+    elevationMax: 993,
+    defaultBounds: [-85.335102, 35.011299, -85.328236, 35.015798],
+    icon: faMountain,
+  },
+  // St. Elmo community
+  {
+    trailName: 'Virginia Ave Greenway',
+    displayName: 'Virginia Ave Greenway',
+    recArea: 'Lookout Mountain',
+    rating: 'easy',
+    color: trailColor('easy', true),
+    distance: 1.4,
+    elevationGain: 70,
+    elevationLoss: 70,
+    elevationMin: 684,
+    elevationMax: 738,
+    defaultBounds: [-85.330467, 34.988378, -85.327892, 35.009049],
+    icon: faRoute,
+  },
+  // Stringers Ridge
+  {
+    trailName: 'Cherokee Trail',
+    displayName: 'Cherokee Trail',
+    recArea: 'Stringers Ridge',
+    rating: 'easy',
+    color: trailColor('easy', false),
+    distance: 1.3,
+    elevationGain: 420,
+    elevationLoss: 195,
+    elevationMin: 780,
+    elevationMax: 1032,
+    defaultBounds: [-85.318451, 35.071312, -85.312099, 35.078617],
+    icon: faMountain,
+  },
+  {
+    trailName: 'Choo Choo',
+    displayName: 'Choo Choo',
+    recArea: 'Stringers Ridge',
+    rating: 'easy',
+    color: trailColor('easy', false),
+    distance: 0.2,
+    elevationGain: 45,
+    elevationLoss: 68,
+    elevationMin: 888,
+    elevationMax: 924,
+    defaultBounds: [-85.315189, 35.071453, -85.314503, 35.072717],
+    icon: faMountain,
+  },
+  {
+    trailName: 'Choo Choo Connector',
+    displayName: 'Choo Choo Connector',
+    recArea: 'Stringers Ridge',
+    rating: 'easy',
+    color: trailColor('easy', false),
+    distance: 0.1,
+    elevationGain: 28,
+    elevationLoss: 40,
+    elevationMin: 881,
+    elevationMax: 909,
+    defaultBounds: [-85.314846, 35.071453, -85.314674, 35.072577],
+    icon: faMountain,
+  },
+  {
+    trailName: 'Double J',
+    displayName: 'Double J',
+    recArea: 'Stringers Ridge',
+    rating: 'intermediate',
+    color: trailColor('intermediate', false),
+    distance: 1.4,
+    elevationGain: 509,
+    elevationLoss: 551,
+    elevationMin: 859,
+    elevationMax: 1009,
+    defaultBounds: [-85.314674, 35.075527, -85.309181, 35.080865],
+    icon: faMountain,
+  },
+  {
+    trailName: 'Gig City',
+    displayName: 'Gig City',
+    recArea: 'Stringers Ridge',
+    rating: 'intermediate',
+    color: trailColor('intermediate', false),
+    distance: 0.6,
+    elevationGain: 220,
+    elevationLoss: 191,
+    elevationMin: 869,
+    elevationMax: 1003,
+    defaultBounds: [-85.317249, 35.072998, -85.313644, 35.075527],
+    icon: faMountain,
+  },
+  {
+    trailName: 'Gig City/Cherokee Trail',
+    displayName: 'Gig City/Cherokee Trail',
+    recArea: 'Stringers Ridge',
+    rating: 'easy',
+    color: trailColor('easy', false),
+    distance: 0.1,
+    elevationGain: 18,
+    elevationLoss: 47,
+    elevationMin: 887,
+    elevationMax: 925,
+    defaultBounds: [-85.314331, 35.073981, -85.313644, 35.075527],
+    icon: faMountain,
+  },
+  {
+    trailName: 'Hill City',
+    displayName: 'Hill City',
+    recArea: 'Stringers Ridge',
+    rating: 'intermediate',
+    color: trailColor('intermediate', false),
+    distance: 1.5,
+    elevationGain: 505,
+    elevationLoss: 438,
+    elevationMin: 759,
+    elevationMax: 956,
+    defaultBounds: [-85.313988, 35.073981, -85.308151, 35.080022],
+    icon: faMountain,
+  },
+  {
+    trailName: 'Spears Ave',
+    displayName: 'Spears Ave',
+    recArea: 'Stringers Ridge',
+    rating: 'easy',
+    color: trailColor('easy', false),
+    distance: 0.1,
+    elevationGain: 39,
+    elevationLoss: 6,
+    elevationMin: 814,
+    elevationMax: 848,
+    defaultBounds: [-85.308838, 35.078477, -85.308151, 35.07932],
+    icon: faMountain,
+  },
+  {
+    trailName: 'Stringers Ridge',
+    displayName: 'Stringers Ridge',
+    recArea: 'Stringers Ridge',
+    rating: '',
+    color: trailColor('', false),
+    distance: 0.3,
+    elevationGain: 61,
+    elevationLoss: 115,
+    elevationMin: 864,
+    elevationMax: 957,
+    defaultBounds: [-85.314503, 35.076089, -85.313129, 35.079179],
+    icon: faMountain,
+  },
+  {
+    trailName: 'Strut',
+    displayName: 'Strut',
+    recArea: 'Stringers Ridge',
+    rating: 'easy',
+    color: trailColor('easy', false),
+    distance: 0.5,
+    elevationGain: 111,
+    elevationLoss: 122,
+    elevationMin: 924,
+    elevationMax: 992,
+    defaultBounds: [-85.312099, 35.078477, -85.30901, 35.080163],
+    icon: faMountain,
+  },
+  {
+    trailName: "White Oak Connector to Stringer's Ridge",
+    displayName: 'White Oak Connector',
+    recArea: 'Stringers Ridge',
+    rating: 'easy',
+    color: trailColor('easy', false),
+    distance: 0.8,
+    elevationGain: 259,
+    elevationLoss: 84,
+    elevationMin: 712,
+    elevationMax: 889,
+    defaultBounds: [-85.314331, 35.079601, -85.308151, 35.086905],
+    icon: faMountain,
+  },
+  // Walden's Ridge Park
+  {
+    trailName: 'Biscuits & Gravy',
+    displayName: 'Biscuits & Gravy',
+    recArea: "Walden's Ridge Park",
+    rating: 'intermediate',
+    color: trailColor('intermediate', false),
+    distance: 0.6,
+    elevationGain: 28,
+    elevationLoss: 333,
+    elevationMin: 1096,
+    elevationMax: 1401,
+    defaultBounds: [-85.314159, 35.138019, -85.310726, 35.139564],
+    icon: faMountain,
+  },
+  {
+    trailName: 'Bread & Butter',
+    displayName: 'Bread & Butter',
+    recArea: "Walden's Ridge Park",
+    rating: 'intermediate',
+    color: trailColor('intermediate', false),
+    distance: 0.4,
+    elevationGain: 8,
+    elevationLoss: 290,
+    elevationMin: 812,
+    elevationMax: 1094,
+    defaultBounds: [-85.310726, 35.136335, -85.307808, 35.138019],
+    icon: faMountain,
+  },
+  {
+    trailName: 'Bullwinkle',
+    displayName: 'Bullwinkle',
+    recArea: "Walden's Ridge Park",
+    rating: 'advanced',
+    color: trailColor('advanced', false),
+    distance: 0.3,
+    elevationGain: 25,
+    elevationLoss: 262,
+    elevationMin: 844,
+    elevationMax: 1081,
+    defaultBounds: [-85.310211, 35.137037, -85.307808, 35.139985],
+    icon: faMountain,
+  },
+  {
+    trailName: 'Clear Eyes',
+    displayName: 'Clear Eyes',
+    recArea: "Walden's Ridge Park",
+    rating: 'advanced',
+    color: trailColor('advanced', false),
+    distance: 0.1,
+    elevationGain: 0,
+    elevationLoss: 57,
+    elevationMin: 1396,
+    elevationMax: 1453,
+    defaultBounds: [-85.315704, 35.137318, -85.315189, 35.137598],
+    icon: faMountain,
+  },
+  {
+    trailName: 'Dale Bermhardt',
+    displayName: 'Dale Bermhardt',
+    recArea: "Walden's Ridge Park",
+    rating: 'advanced',
+    color: trailColor('advanced', false),
+    distance: 0.5,
+    elevationGain: 32,
+    elevationLoss: 342,
+    elevationMin: 1092,
+    elevationMax: 1402,
+    defaultBounds: [-85.313301, 35.139704, -85.309868, 35.141108],
+    icon: faMountain,
+  },
+  {
+    trailName: 'DeVito',
+    displayName: 'DeVito',
+    recArea: "Walden's Ridge Park",
+    rating: 'advanced',
+    color: trailColor('advanced', false),
+    distance: 0.0,
+    elevationGain: 0,
+    elevationLoss: 45,
+    elevationMin: 1402,
+    elevationMax: 1447,
+    defaultBounds: [-85.313988, 35.141108, -85.313301, 35.141108],
+    icon: faMountain,
+  },
+  {
+    trailName: 'Escalator',
+    displayName: 'Escalator',
+    recArea: "Walden's Ridge Park",
+    rating: 'easy',
+    color: trailColor('easy', false),
+    distance: 2.4,
+    elevationGain: 227,
+    elevationLoss: 968,
+    elevationMin: 745,
+    elevationMax: 1486,
+    defaultBounds: [-85.316391, 35.132685, -85.30695, 35.137879],
+    icon: faMountain,
+  },
+  {
+    trailName: 'Falls Connector',
+    displayName: 'Falls Connector',
+    recArea: "Walden's Ridge Park",
+    rating: '',
+    color: trailColor('', false),
+    distance: 0.1,
+    elevationGain: 32,
+    elevationLoss: 72,
+    elevationMin: 1412,
+    elevationMax: 1469,
+    defaultBounds: [-85.313988, 35.141669, -85.312958, 35.143213],
+    icon: faMountain,
+  },
+  {
+    trailName: 'Gambler',
+    displayName: 'Gambler',
+    recArea: "Walden's Ridge Park",
+    rating: 'advanced',
+    color: trailColor('advanced', false),
+    distance: 0.4,
+    elevationGain: 12,
+    elevationLoss: 306,
+    elevationMin: 1096,
+    elevationMax: 1390,
+    defaultBounds: [-85.314846, 35.137458, -85.310898, 35.1383],
+    icon: faMountain,
+  },
+  {
+    trailName: 'Lower Lift Ticket',
+    displayName: 'Lower Lift Ticket',
+    recArea: "Walden's Ridge Park",
+    rating: 'easy',
+    color: trailColor('easy', false),
+    distance: 1.1,
+    elevationGain: 470,
+    elevationLoss: 195,
+    elevationMin: 804,
+    elevationMax: 1094,
+    defaultBounds: [-85.309353, 35.136896, -85.30695, 35.142933],
+    icon: faMountain,
+  },
+  {
+    trailName: 'Middle Traverse',
+    displayName: 'Middle Traverse',
+    recArea: "Walden's Ridge Park",
+    rating: 'intermediate',
+    color: trailColor('intermediate', false),
+    distance: 0.4,
+    elevationGain: 117,
+    elevationLoss: 86,
+    elevationMin: 1083,
+    elevationMax: 1119,
+    defaultBounds: [-85.311756, 35.137037, -85.309353, 35.14181],
+    icon: faMountain,
+  },
+  {
+    trailName: 'Ollie',
+    displayName: 'Ollie',
+    recArea: "Walden's Ridge Park",
+    rating: 'intermediate',
+    color: trailColor('intermediate', false),
+    distance: 0.4,
+    elevationGain: 25,
+    elevationLoss: 278,
+    elevationMin: 863,
+    elevationMax: 1118,
+    defaultBounds: [-85.309525, 35.139564, -85.307465, 35.141248],
+    icon: faMountain,
+  },
+  {
+    trailName: 'Pocket Rockets',
+    displayName: 'Pocket Rockets',
+    recArea: "Walden's Ridge Park",
+    rating: 'advanced',
+    color: trailColor('advanced', false),
+    distance: 0.4,
+    elevationGain: 31,
+    elevationLoss: 309,
+    elevationMin: 823,
+    elevationMax: 1102,
+    defaultBounds: [-85.311241, 35.136054, -85.307636, 35.137318],
+    icon: faMountain,
+  },
+  {
+    trailName: 'Stone Mason',
+    displayName: 'Stone Mason',
+    recArea: "Walden's Ridge Park",
+    rating: 'expert',
+    color: trailColor('expert', false),
+    distance: 0.1,
+    elevationGain: 0,
+    elevationLoss: 125,
+    elevationMin: 1389,
+    elevationMax: 1514,
+    defaultBounds: [-85.316048, 35.137598, -85.314846, 35.1383],
+    icon: faMountain,
+  },
+  {
+    trailName: 'Thrasher',
+    displayName: 'Thrasher',
+    recArea: "Walden's Ridge Park",
+    rating: 'intermediate',
+    color: trailColor('intermediate', false),
+    distance: 0.4,
+    elevationGain: 9,
+    elevationLoss: 284,
+    elevationMin: 1115,
+    elevationMax: 1390,
+    defaultBounds: [-85.312786, 35.140827, -85.309525, 35.14209],
+    icon: faMountain,
+  },
+  {
+    trailName: 'Upper Lift Ticket',
+    displayName: 'Upper Lift Ticket',
+    recArea: "Walden's Ridge Park",
+    rating: 'advanced',
+    color: trailColor('advanced', false),
+    distance: 0.7,
+    elevationGain: 128,
+    elevationLoss: 426,
+    elevationMin: 1088,
+    elevationMax: 1388,
+    defaultBounds: [-85.312614, 35.14181, -85.309353, 35.144056],
+    icon: faMountain,
+  },
+  {
+    trailName: 'Upper Traverse',
+    displayName: 'Upper Traverse',
+    recArea: "Walden's Ridge Park",
+    rating: 'intermediate',
+    color: trailColor('intermediate', false),
+    distance: 0.4,
+    elevationGain: 84,
+    elevationLoss: 90,
+    elevationMin: 1386,
+    elevationMax: 1417,
+    defaultBounds: [-85.315533, 35.136896, -85.312614, 35.14195],
+    icon: faMountain,
+  },
+  {
+    trailName: 'W Trail',
+    displayName: 'W Trail',
+    recArea: "Walden's Ridge Park",
+    rating: 'intermediate',
+    color: trailColor('intermediate', false),
+    distance: 0.5,
+    elevationGain: 75,
+    elevationLoss: 187,
+    elevationMin: 1390,
+    elevationMax: 1525,
+    defaultBounds: [-85.316734, 35.137318, -85.312958, 35.141669],
+    icon: faMountain,
+  },
+  {
+    trailName: 'Walden Falls',
+    displayName: 'Walden Falls',
+    recArea: "Walden's Ridge Park",
+    rating: '',
+    color: trailColor('', false),
+    distance: 0.3,
+    elevationGain: 147,
+    elevationLoss: 120,
+    elevationMin: 1301,
+    elevationMax: 1421,
+    defaultBounds: [-85.312958, 35.143213, -85.311584, 35.145459],
+    icon: faMountain,
+  },
+  {
+    trailName: 'Waldens Godsey Connector',
+    displayName: "Walden's Godsey Connector",
+    recArea: "Walden's Ridge Park",
+    rating: '',
+    color: trailColor('', false),
+    distance: 1.4,
+    elevationGain: 121,
+    elevationLoss: 156,
+    elevationMin: 699,
+    elevationMax: 815,
+    defaultBounds: [-85.306971, 35.123588, -85.300156, 35.136076],
+    icon: faMountain,
+  },
+  // White Oak Mountain
+  {
+    trailName: 'Barred Owl',
+    displayName: 'Barred Owl',
+    recArea: 'White Oak Mountain',
+    rating: 'easy',
+    color: trailColor('easy', false),
+    distance: 0.8,
+    elevationGain: 211,
+    elevationLoss: 236,
+    elevationMin: 971,
+    elevationMax: 1031,
+    defaultBounds: [-85.063705, 35.0411, -85.058727, 35.045176],
+    icon: faMountain,
+  },
+  {
+    trailName: 'Big Red',
+    displayName: 'Big Red',
+    recArea: 'White Oak Mountain',
+    rating: 'intermediate',
+    color: trailColor('intermediate', false),
+    distance: 1.2,
+    elevationGain: 474,
+    elevationLoss: 316,
+    elevationMin: 877,
+    elevationMax: 1151,
+    defaultBounds: [-85.06422, 35.041662, -85.056839, 35.044052],
+    icon: faMountain,
+  },
+  {
+    trailName: 'Black Bear',
+    displayName: 'Black Bear',
+    recArea: 'White Oak Mountain',
+    rating: 'intermediate',
+    color: trailColor('intermediate', false),
+    distance: 0.2,
+    elevationGain: 17,
+    elevationLoss: 79,
+    elevationMin: 906,
+    elevationMax: 968,
+    defaultBounds: [-85.059757, 35.049814, -85.057526, 35.050938],
+    icon: faMountain,
+  },
+  {
+    trailName: 'Bobcat',
+    displayName: 'Bobcat',
+    recArea: 'White Oak Mountain',
+    rating: 'easy',
+    color: trailColor('easy', false),
+    distance: 0.5,
+    elevationGain: 134,
+    elevationLoss: 95,
+    elevationMin: 975,
+    elevationMax: 1047,
+    defaultBounds: [-85.06422, 35.041241, -85.059071, 35.042225],
+    icon: faMountain,
+  },
+  {
+    trailName: 'Christmas Top',
+    displayName: 'Christmas Top',
+    recArea: 'White Oak Mountain',
+    rating: '',
+    color: trailColor('', false),
+    distance: 0.7,
+    elevationGain: 468,
+    elevationLoss: 91,
+    elevationMin: 871,
+    elevationMax: 1248,
+    defaultBounds: [-85.066624, 35.042084, -85.056496, 35.043068],
+    icon: faMountain,
+  },
+  {
+    trailName: 'Copperhead Trail',
+    displayName: 'Copperhead Trail',
+    recArea: 'White Oak Mountain',
+    rating: 'intermediate',
+    color: trailColor('intermediate', false),
+    distance: 0.2,
+    elevationGain: 25,
+    elevationLoss: 155,
+    elevationMin: 1038,
+    elevationMax: 1168,
+    defaultBounds: [-85.060959, 35.05417, -85.059242, 35.055997],
+    icon: faMountain,
+  },
+  {
+    trailName: 'Coyote',
+    displayName: 'Coyote',
+    recArea: 'White Oak Mountain',
+    rating: 'easy',
+    color: trailColor('easy', false),
+    distance: 1.0,
+    elevationGain: 260,
+    elevationLoss: 268,
+    elevationMin: 974,
+    elevationMax: 1041,
+    defaultBounds: [-85.061302, 35.045738, -85.055637, 35.049252],
+    icon: faMountain,
+  },
+  {
+    trailName: 'Dogwood',
+    displayName: 'Dogwood',
+    recArea: 'White Oak Mountain',
+    rating: 'intermediate',
+    color: trailColor('intermediate', false),
+    distance: 1.1,
+    elevationGain: 171,
+    elevationLoss: 194,
+    elevationMin: 863,
+    elevationMax: 996,
+    defaultBounds: [-85.061474, 35.042927, -85.054264, 35.050235],
+    icon: faMountain,
+  },
+  {
+    trailName: 'Forest',
+    displayName: 'Forest',
+    recArea: 'White Oak Mountain',
+    rating: 'intermediate',
+    color: trailColor('intermediate', false),
+    distance: 0.7,
+    elevationGain: 257,
+    elevationLoss: 161,
+    elevationMin: 998,
+    elevationMax: 1164,
+    defaultBounds: [-85.066967, 35.03604, -85.059757, 35.040679],
+    icon: faMountain,
+  },
+  {
+    trailName: 'Give Me A Break',
+    displayName: 'Give Me A Break',
+    recArea: 'White Oak Mountain',
+    rating: 'expert',
+    color: trailColor('expert', false),
+    distance: 0.1,
+    elevationGain: 235,
+    elevationLoss: 5,
+    elevationMin: 991,
+    elevationMax: 1226,
+    defaultBounds: [-85.063534, 35.048268, -85.061131, 35.04883],
+    icon: faMountain,
+  },
+  {
+    trailName: 'Krebs Cycle',
+    displayName: 'Krebs Cycle',
+    recArea: 'White Oak Mountain',
+    rating: 'intermediate',
+    color: trailColor('intermediate', false),
+    distance: 1.1,
+    elevationGain: 354,
+    elevationLoss: 278,
+    elevationMin: 897,
+    elevationMax: 1085,
+    defaultBounds: [-85.061131, 35.044192, -85.056496, 35.051219],
+    icon: faMountain,
+  },
+  {
+    trailName: 'Magnolia',
+    displayName: 'Magnolia',
+    recArea: 'White Oak Mountain',
+    rating: '',
+    color: trailColor('', false),
+    distance: 0.1,
+    elevationGain: 3,
+    elevationLoss: 54,
+    elevationMin: 1059,
+    elevationMax: 1112,
+    defaultBounds: [-85.064392, 35.039273, -85.064392, 35.040257],
+    icon: faMountain,
+  },
+  {
+    trailName: 'Mid Ridge',
+    displayName: 'Mid Ridge',
+    recArea: 'White Oak Mountain',
+    rating: 'intermediate',
+    color: trailColor('intermediate', false),
+    distance: 1.2,
+    elevationGain: 332,
+    elevationLoss: 211,
+    elevationMin: 973,
+    elevationMax: 1099,
+    defaultBounds: [-85.066967, 35.03857, -85.059586, 35.051219],
+    icon: faMountain,
+  },
+  {
+    trailName: 'North Nuthatch',
+    displayName: 'North Nuthatch',
+    recArea: 'White Oak Mountain',
+    rating: '',
+    color: trailColor('', false),
+    distance: 0.1,
+    elevationGain: 83,
+    elevationLoss: 63,
+    elevationMin: 931,
+    elevationMax: 1014,
+    defaultBounds: [-85.060101, 35.039554, -85.059242, 35.0411],
+    icon: faMountain,
+  },
+  {
+    trailName: 'North Side Climb Trail',
+    displayName: 'North Side Climb Trail',
+    recArea: 'White Oak Mountain',
+    rating: 'intermediate',
+    color: trailColor('intermediate', false),
+    distance: 1.1,
+    elevationGain: 471,
+    elevationLoss: 165,
+    elevationMin: 794,
+    elevationMax: 1115,
+    defaultBounds: [-85.057697, 35.050095, -85.053749, 35.054732],
+    icon: faMountain,
+  },
+  {
+    trailName: 'North Tower Fire Road',
+    displayName: 'North Tower Fire Road',
+    recArea: 'White Oak Mountain',
+    rating: '',
+    color: trailColor('', false),
+    distance: 0.3,
+    elevationGain: 63,
+    elevationLoss: 142,
+    elevationMin: 975,
+    elevationMax: 1115,
+    defaultBounds: [-85.059586, 35.050797, -85.055981, 35.051641],
+    icon: faMountain,
+  },
+  {
+    trailName: 'Nutty Buddy (Upper)',
+    displayName: 'Nutty Buddy (Upper)',
+    recArea: 'White Oak Mountain',
+    rating: 'advanced',
+    color: trailColor('advanced', false),
+    distance: 0.4,
+    elevationGain: 51,
+    elevationLoss: 188,
+    elevationMin: 979,
+    elevationMax: 1118,
+    defaultBounds: [-85.059414, 35.050797, -85.055981, 35.052343],
+    icon: faMountain,
+  },
+  {
+    trailName: 'Picnic',
+    displayName: 'Picnic',
+    recArea: 'White Oak Mountain',
+    rating: '',
+    color: trailColor('', false),
+    distance: 0.1,
+    elevationGain: 64,
+    elevationLoss: 44,
+    elevationMin: 1046,
+    elevationMax: 1107,
+    defaultBounds: [-85.058899, 35.046019, -85.057354, 35.047565],
+    icon: faMountain,
+  },
+  {
+    trailName: 'Pine',
+    displayName: 'Pine',
+    recArea: 'White Oak Mountain',
+    rating: 'easy',
+    color: trailColor('easy', false),
+    distance: 1.1,
+    elevationGain: 370,
+    elevationLoss: 172,
+    elevationMin: 870,
+    elevationMax: 1069,
+    defaultBounds: [-85.064907, 35.041522, -85.055809, 35.043068],
+    icon: faMountain,
+  },
+  {
+    trailName: 'Points Unknown',
+    displayName: 'Points Unknown',
+    recArea: 'White Oak Mountain',
+    rating: 'intermediate',
+    color: trailColor('intermediate', false),
+    distance: 1.7,
+    elevationGain: 419,
+    elevationLoss: 492,
+    elevationMin: 1035,
+    elevationMax: 1203,
+    defaultBounds: [-85.067997, 35.038851, -85.059242, 35.055294],
+    icon: faMountain,
+  },
+  {
+    trailName: 'Rolling Green',
+    displayName: 'Rolling Green',
+    recArea: 'White Oak Mountain',
+    rating: 'easy',
+    color: trailColor('easy', false),
+    distance: 1.8,
+    elevationGain: 420,
+    elevationLoss: 353,
+    elevationMin: 871,
+    elevationMax: 1034,
+    defaultBounds: [-85.063019, 35.043068, -85.056152, 35.048408],
+    icon: faMountain,
+  },
+  {
+    trailName: 'South Nuthatch',
+    displayName: 'South Nuthatch',
+    recArea: 'White Oak Mountain',
+    rating: '',
+    color: trailColor('', false),
+    distance: 0.2,
+    elevationGain: 72,
+    elevationLoss: 76,
+    elevationMin: 975,
+    elevationMax: 1046,
+    defaultBounds: [-85.061646, 35.037446, -85.059757, 35.03857],
+    icon: faMountain,
+  },
+  {
+    trailName: 'Southern Mission',
+    displayName: 'Southern Mission',
+    recArea: 'White Oak Mountain',
+    rating: 'intermediate',
+    color: trailColor('intermediate', false),
+    distance: 2.0,
+    elevationGain: 515,
+    elevationLoss: 465,
+    elevationMin: 1010,
+    elevationMax: 1130,
+    defaultBounds: [-85.065079, 35.03604, -85.059071, 35.041241],
+    icon: faMountain,
+  },
+  {
+    trailName: 'Summer',
+    displayName: 'Summer',
+    recArea: 'White Oak Mountain',
+    rating: '',
+    color: trailColor('', false),
+    distance: 0.2,
+    elevationGain: 21,
+    elevationLoss: 70,
+    elevationMin: 970,
+    elevationMax: 1025,
+    defaultBounds: [-85.063019, 35.037446, -85.059757, 35.038851],
+    icon: faMountain,
+  },
+  {
+    trailName: 'The Fall',
+    displayName: 'The Fall',
+    recArea: 'White Oak Mountain',
+    rating: 'advanced',
+    color: trailColor('advanced', false),
+    distance: 0.2,
+    elevationGain: 141,
+    elevationLoss: 4,
+    elevationMin: 911,
+    elevationMax: 1048,
+    defaultBounds: [-85.059757, 35.045035, -85.059586, 35.047565],
+    icon: faMountain,
+  },
+  {
+    trailName: 'Tower',
+    displayName: 'Tower',
+    recArea: 'White Oak Mountain',
+    rating: 'intermediate',
+    color: trailColor('intermediate', false),
+    distance: 0.4,
+    elevationGain: 74,
+    elevationLoss: 146,
+    elevationMin: 901,
+    elevationMax: 1002,
+    defaultBounds: [-85.059586, 35.049533, -85.055122, 35.051219],
+    icon: faMountain,
+  },
+  {
+    trailName: 'Upper Copperhead',
+    displayName: 'Upper Copperhead',
+    recArea: 'White Oak Mountain',
+    rating: 'intermediate',
+    color: trailColor('intermediate', false),
+    distance: 0.3,
+    elevationGain: 48,
+    elevationLoss: 104,
+    elevationMin: 967,
+    elevationMax: 1032,
+    defaultBounds: [-85.059586, 35.051219, -85.059242, 35.055294],
+    icon: faMountain,
+  },
+  {
+    trailName: 'White Oak Ridge',
+    displayName: 'White Oak Ridge',
+    recArea: 'White Oak Mountain',
+    rating: 'expert',
+    color: trailColor('expert', false),
+    distance: 1.6,
+    elevationGain: 381,
+    elevationLoss: 366,
+    elevationMin: 1117,
+    elevationMax: 1259,
+    defaultBounds: [-85.067997, 35.038851, -85.060959, 35.05417],
+    icon: faMountain,
+  },
+  // White Oak Park
+  {
+    trailName: 'White Oak Park Trail',
+    displayName: 'White Oak Park Trail',
+    recArea: 'Stringers Ridge',
+    rating: '',
+    color: trailColor('', false),
+    distance: 0.6,
+    elevationGain: 141,
+    elevationLoss: 53,
+    elevationMin: 705,
+    elevationMax: 798,
+    defaultBounds: [-85.308151, 35.085641, -85.301456, 35.087186],
+    icon: faMountain,
+  },
+  // Godsey Ridge
+  {
+    trailName: 'Godsey Ridge Green',
+    displayName: 'Godsey Ridge Green',
+    recArea: 'Godsey Ridge',
+    rating: 'easy',
+    color: trailColor('easy', false),
+    distance: 0.3,
+    elevationGain: 60,
+    elevationLoss: 56,
+    elevationMin: 774,
+    elevationMax: 803,
+    defaultBounds: [-85.302469, 35.121044, -85.300579, 35.123595],
+    icon: faMountain,
+  },
+  {
+    trailName: 'Godsey Ridge Blue 1',
+    displayName: 'Godsey Ridge Blue 1',
+    recArea: 'Godsey Ridge',
+    rating: 'intermediate',
+    color: trailColor('intermediate', false),
+    distance: 0.2,
+    elevationGain: 95,
+    elevationLoss: 151,
+    elevationMin: 776,
+    elevationMax: 881,
+    defaultBounds: [-85.302859, 35.120159, -85.301371, 35.121293],
+    icon: faMountain,
+  },
+  {
+    trailName: 'Godsey Ridge Blue 2',
+    displayName: 'Godsey Ridge Blue 2',
+    recArea: 'Godsey Ridge',
+    rating: 'intermediate',
+    color: trailColor('intermediate', false),
+    distance: 0.2,
+    elevationGain: 114,
+    elevationLoss: 59,
+    elevationMin: 776,
+    elevationMax: 841,
+    defaultBounds: [-85.301711, 35.120772, -85.300742, 35.123043],
+    icon: faMountain,
+  },
+  {
+    trailName: 'Godsey Ridge Expert Spur',
+    displayName: 'Godsey Ridge Expert Spur',
+    recArea: 'Godsey Ridge',
+    rating: 'expert',
+    color: trailColor('expert', false),
+    distance: 0.0,
+    elevationGain: 16,
+    elevationLoss: 12,
+    elevationMin: 940,
+    elevationMax: 952,
+    defaultBounds: [-85.300538, 35.12061, -85.300249, 35.120873],
+    icon: faMountain,
+  },
+  {
+    trailName: 'Godsey Ridge Expert 1',
+    displayName: 'Godsey Ridge Expert 1',
+    recArea: 'Godsey Ridge',
+    rating: 'expert',
+    color: trailColor('expert', false),
+    distance: 0.3,
+    elevationGain: 211,
+    elevationLoss: 100,
+    elevationMin: 832,
+    elevationMax: 949,
+    defaultBounds: [-85.302012, 35.120035, -85.30031, 35.121158],
+    icon: faMountain,
+  },
+  {
+    trailName: 'Godsey Ridge Expert 2',
+    displayName: 'Godsey Ridge Expert 2',
+    recArea: 'Godsey Ridge',
+    rating: 'expert',
+    color: trailColor('expert', false),
+    distance: 0.3,
+    elevationGain: 90,
+    elevationLoss: 201,
+    elevationMin: 831,
+    elevationMax: 943,
+    defaultBounds: [-85.301431, 35.120873, -85.299851, 35.122439],
+    icon: faMountain,
+  },
+];
